fix(dashboard): rename auth.js to auth.svelte.js for rune support
$state() rune only works in .svelte and .svelte.js files, not plain
.js. The auth module was failing silently at runtime, leaving the
app div empty.
