+++ assets/admin/assets/index-BabJX8K1.js
@@ -1,0 +1,13 @@
+var $n=Object.defineProperty;var pa=t=>{throw TypeError(t)};var Rn=(t,e,r)=>e in t?$n(t,e,{enumerable:!0,configurable:!0,writable:!0,value:r}):t[e]=r;var Vt=(t,e,r)=>Rn(t,typeof e!="symbol"?e+"":e,r),Or=(t,e,r)=>e.has(t)||pa("Cannot "+r);var c=(t,e,r)=>(Or(t,e,"read from private field"),r?r.call(t):e.get(t)),K=(t,e,r)=>e.has(t)?pa("Cannot add the same private member more than once"):e instanceof WeakSet?e.add(t):e.set(t,r),Y=(t,e,r,a)=>(Or(t,e,"write to private field"),a?a.call(t,r):e.set(t,r),r),et=(t,e,r)=>(Or(t,e,"access private method"),r);(function(){const e=document.createElement("link").relList;if(e&&e.supports&&e.supports("modulepreload"))return;for(const n of document.querySelectorAll('link[rel="modulepreload"]'))a(n);new MutationObserver(n=>{for(const s of n)if(s.type==="childList")for(const o of s.addedNodes)o.tagName==="LINK"&&o.rel==="modulepreload"&&a(o)}).observe(document,{childList:!0,subtree:!0});function r(n){const s={};return n.integrity&&(s.integrity=n.integrity),n.referrerPolicy&&(s.referrerPolicy=n.referrerPolicy),n.crossOrigin==="use-credentials"?s.credentials="include":n.crossOrigin==="anonymous"?s.credentials="omit":s.credentials="same-origin",s}function a(n){if(n.ep)return;n.ep=!0;const s=r(n);fetch(n.href,s)}})();const Dn="5";var Ta;typeof window<"u"&&((Ta=window.__svelte??(window.__svelte={})).v??(Ta.v=new Set)).add(Dn);const On=1,Ln=2,Ca=4,Pn=8,Fn=16,jn=1,Bn=2,ut=Symbol(),Ma="http://www.w3.org/1999/xhtml",Un=!1;var $a=Array.isArray,zn=Array.prototype.indexOf,qe=Array.prototype.includes,Mr=Array.from,Hn=Object.defineProperty,ar=Object.getOwnPropertyDescriptor,qn=Object.getOwnPropertyDescriptors,Vn=Object.prototype,Wn=Array.prototype,Ra=Object.getPrototypeOf,ha=Object.isExtensible;const Ft=()=>{};function Jn(t){for(var e=0;e<t.length;e++)t[e]()}function Da(){var t,e,r=new Promise((a,n)=>{t=a,e=n});return{promise:r,resolve:t,reject:e}}function Oa(t,e){if(Array.isArray(t))return t;if(!(Symbol.iterator in t))return Array.from(t);const r=[];for(const a of t)if(r.push(a),r.length===e)break;return r}const _t=2,ir=4,$r=8,La=1<<24,re=16,Gt=32,ge=64,zr=128,jt=512,ft=1024,ht=2048,ae=4096,yt=8192,Bt=16384,Pe=32768,_a=1<<25,Ve=65536,Hr=1<<17,Kn=1<<18,Xe=1<<19,Gn=1<<20,ee=1<<25,De=65536,qr=1<<21,or=1<<22,pe=1<<23,Lr=Symbol("$state"),se=new class extends Error{constructor(){super(...arguments);Vt(this,"name","StaleReactionError");Vt(this,"message","The reaction that called `getAbortSignal()` was re-run or destroyed")}};var Ia;const Yn=!!((Ia=globalThis.document)!=null&&Ia.contentType)&&globalThis.document.contentType.includes("xml");function Zn(){throw new Error("https://svelte.dev/e/async_derived_orphan")}function Xn(t,e,r){throw new Error("https://svelte.dev/e/each_key_duplicate")}function Qn(t){throw new Error("https://svelte.dev/e/effect_in_teardown")}function ts(){throw new Error("https://svelte.dev/e/effect_in_unowned_derived")}function es(t){throw new Error("https://svelte.dev/e/effect_orphan")}function rs(){throw new Error("https://svelte.dev/e/effect_update_depth_exceeded")}function as(){throw new Error("https://svelte.dev/e/state_descriptors_fixed")}function ns(){throw new Error("https://svelte.dev/e/state_prototype_fixed")}function ss(){throw new Error("https://svelte.dev/e/state_unsafe_mutation")}function is(){throw new Error("https://svelte.dev/e/svelte_boundary_reset_onerror")}function os(){console.warn("https://svelte.dev/e/derived_inert")}function ls(){console.warn("https://svelte.dev/e/svelte_boundary_reset_noop")}function Pa(t){return t===this.v}function ds(t,e){return t!=t?e==e:t!==e||t!==null&&typeof t=="object"||typeof t=="function"}function Fa(t){return!ds(t,this.v)}let fs=!1,zt=null;function We(t){zt=t}function At(t,e=!1,r){zt={p:zt,i:!1,c:null,e:null,s:t,x:null,r:X,l:null}}function Tt(t){var e=zt,r=e.e;if(r!==null){e.e=null;for(var a of r)nn(a)}return e.i=!0,zt=e.p,{}}function ja(){return!0}let ke=[];function Ba(){var t=ke;ke=[],Jn(t)}function he(t){if(ke.length===0&&!nr){var e=ke;queueMicrotask(()=>{e===ke&&Ba()})}ke.push(t)}function us(){for(;ke.length>0;)Ba()}function Ua(t){var e=X;if(e===null)return Z.f|=pe,t;if((e.f&Pe)===0&&(e.f&ir)===0)throw t;ve(t,e)}function ve(t,e){for(;e!==null;){if((e.f&zr)!==0){if((e.f&Pe)===0)throw t;try{e.b.error(t);return}catch(r){t=r}}e=e.parent}throw t}const cs=-7169;function ot(t,e){t.f=t.f&cs|e}function ea(t){(t.f&jt)!==0||t.deps===null?ot(t,ft):ot(t,ae)}function za(t){if(t!==null)for(const e of t)(e.f&_t)===0||(e.f&De)===0||(e.f^=De,za(e.deps))}function Ha(t,e,r){(t.f&ht)!==0?e.add(t):(t.f&ae)!==0&&r.add(t),za(t.deps),ot(t,ft)}const we=new Set;let z=null,vt=null,Vr=null,nr=!1,Pr=!1,Fe=null,yr=null;var ga=0;let vs=1;var je,Be,Ae,ie,Xt,dr,Nt,fr,ue,oe,Qt,Ue,ze,Te,dt,br,qa,mr,Wr,wr,ps;const Ir=class Ir{constructor(){K(this,dt);Vt(this,"id",vs++);Vt(this,"current",new Map);Vt(this,"previous",new Map);K(this,je,new Set);K(this,Be,new Set);K(this,Ae,new Set);K(this,ie,new Map);K(this,Xt,new Map);K(this,dr,null);K(this,Nt,[]);K(this,fr,[]);K(this,ue,new Set);K(this,oe,new Set);K(this,Qt,new Map);K(this,Ue,new Set);Vt(this,"is_fork",!1);K(this,ze,!1);K(this,Te,new Set)}skip_effect(e){c(this,Qt).has(e)||c(this,Qt).set(e,{d:[],m:[]}),c(this,Ue).delete(e)}unskip_effect(e,r=a=>this.schedule(a)){var a=c(this,Qt).get(e);if(a){c(this,Qt).delete(e);for(var n of a.d)ot(n,ht),r(n);for(n of a.m)ot(n,ae),r(n)}c(this,Ue).add(e)}capture(e,r,a=!1){e.v!==ut&&!this.previous.has(e)&&this.previous.set(e,e.v),(e.f&pe)===0&&(this.current.set(e,[r,a]),vt==null||vt.set(e,r)),this.is_fork||(e.v=r)}activate(){z=this}deactivate(){z=null,vt=null}flush(){try{Pr=!0,z=this,et(this,dt,mr).call(this)}finally{ga=0,Vr=null,Fe=null,yr=null,Pr=!1,z=null,vt=null,Me.clear()}}discard(){for(const e of c(this,Be))e(this);c(this,Be).clear(),c(this,Ae).clear(),we.delete(this)}register_created_effect(e){c(this,fr).push(e)}increment(e,r){let a=c(this,ie).get(r)??0;if(c(this,ie).set(r,a+1),e){let n=c(this,Xt).get(r)??0;c(this,Xt).set(r,n+1)}}decrement(e,r,a){let n=c(this,ie).get(r)??0;if(n===1?c(this,ie).delete(r):c(this,ie).set(r,n-1),e){let s=c(this,Xt).get(r)??0;s===1?c(this,Xt).delete(r):c(this,Xt).set(r,s-1)}c(this,ze)||a||(Y(this,ze,!0),he(()=>{Y(this,ze,!1),this.flush()}))}transfer_effects(e,r){for(const a of e)c(this,ue).add(a);for(const a of r)c(this,oe).add(a);e.clear(),r.clear()}oncommit(e){c(this,je).add(e)}ondiscard(e){c(this,Be).add(e)}on_fork_commit(e){c(this,Ae).add(e)}run_fork_commit_callbacks(){for(const e of c(this,Ae))e(this);c(this,Ae).clear()}settled(){return(c(this,dr)??Y(this,dr,Da())).promise}static ensure(){if(z===null){const e=z=new Ir;Pr||(we.add(z),nr||he(()=>{z===e&&e.flush()}))}return z}apply(){{vt=null;return}}schedule(e){var n;if(Vr=e,(n=e.b)!=null&&n.is_pending&&(e.f&(ir|$r|La))!==0&&(e.f&Pe)===0){e.b.defer_effect(e);return}for(var r=e;r.parent!==null;){r=r.parent;var a=r.f;if(Fe!==null&&r===X&&(Z===null||(Z.f&_t)===0))return;if((a&(ge|Gt))!==0){if((a&ft)===0)return;r.f^=ft}}c(this,Nt).push(r)}};je=new WeakMap,Be=new WeakMap,Ae=new WeakMap,ie=new WeakMap,Xt=new WeakMap,dr=new WeakMap,Nt=new WeakMap,fr=new WeakMap,ue=new WeakMap,oe=new WeakMap,Qt=new WeakMap,Ue=new WeakMap,ze=new WeakMap,Te=new WeakMap,dt=new WeakSet,br=function(){return this.is_fork||c(this,Xt).size>0},qa=function(){for(const a of c(this,Te))for(const n of c(a,Xt).keys()){for(var e=!1,r=n;r.parent!==null;){if(c(this,Qt).has(r)){e=!0;break}r=r.parent}if(!e)return!0}return!1},mr=function(){var u,d;if(ga++>1e3&&(we.delete(this),_s()),!et(this,dt,br).call(this)){for(const f of c(this,ue))c(this,oe).delete(f),ot(f,ht),this.schedule(f);for(const f of c(this,oe))ot(f,ae),this.schedule(f)}const e=c(this,Nt);Y(this,Nt,[]),this.apply();var r=Fe=[],a=[],n=yr=[];for(const f of e)try{et(this,dt,Wr).call(this,f,r,a)}catch(_){throw Ja(f),_}if(z=null,n.length>0){var s=Ir.ensure();for(const f of n)s.schedule(f)}if(Fe=null,yr=null,et(this,dt,br).call(this)||et(this,dt,qa).call(this)){et(this,dt,wr).call(this,a),et(this,dt,wr).call(this,r);for(const[f,_]of c(this,Qt))Wa(f,_)}else{c(this,ie).size===0&&we.delete(this),c(this,ue).clear(),c(this,oe).clear();for(const f of c(this,je))f(this);c(this,je).clear(),xa(a),xa(r),(u=c(this,dr))==null||u.resolve()}var o=z;if(c(this,Nt).length>0){const f=o??(o=this);c(f,Nt).push(...c(this,Nt).filter(_=>!c(f,Nt).includes(_)))}o!==null&&(we.add(o),et(d=o,dt,mr).call(d))},Wr=function(e,r,a){e.f^=ft;for(var n=e.first;n!==null;){var s=n.f,o=(s&(Gt|ge))!==0,u=o&&(s&ft)!==0,d=u||(s&yt)!==0||c(this,Qt).has(n);if(!d&&n.fn!==null){o?n.f^=ft:(s&ir)!==0?r.push(n):_r(n)&&((s&re)!==0&&c(this,oe).add(n),Ge(n));var f=n.first;if(f!==null){n=f;continue}}for(;n!==null;){var _=n.next;if(_!==null){n=_;break}n=n.parent}}},wr=function(e){for(var r=0;r<e.length;r+=1)Ha(e[r],c(this,ue),c(this,oe))},ps=function(){var _,b,y;for(const g of we){var e=g.id<this.id,r=[];for(const[v,[w,h]]of this.current){if(g.current.has(v)){var a=g.current.get(v)[0];if(e&&w!==a)g.current.set(v,[w,h]);else continue}r.push(v)}var n=[...g.current.keys()].filter(v=>!this.current.has(v));if(n.length===0)e&&g.discard();else if(r.length>0){if(e)for(const v of c(this,Ue))g.unskip_effect(v,w=>{var h;(w.f&(re|or))!==0?g.schedule(w):et(h=g,dt,wr).call(h,[w])});g.activate();var s=new Set,o=new Map;for(var u of r)Va(u,n,s,o);o=new Map;var d=[...g.current.keys()].filter(v=>this.current.has(v)?this.current.get(v)[0]!==v:!0);for(const v of c(this,fr))(v.f&(Bt|yt|Hr))===0&&ra(v,d,o)&&((v.f&(or|re))!==0?(ot(v,ht),g.schedule(v)):c(g,ue).add(v));if(c(g,Nt).length>0){g.apply();for(var f of c(g,Nt))et(_=g,dt,Wr).call(_,f,[],[]);Y(g,Nt,[])}g.deactivate()}}for(const g of we)c(g,Te).has(this)&&(c(g,Te).delete(this),c(g,Te).size===0&&!et(b=g,dt,br).call(b)&&(g.activate(),et(y=g,dt,mr).call(y)))};let Oe=Ir;function hs(t){var e=nr;nr=!0;try{for(var r;;){if(us(),z===null)return r;z.flush()}}finally{nr=e}}function _s(){try{rs()}catch(t){ve(t,Vr)}}let Wt=null;function xa(t){var e=t.length;if(e!==0){for(var r=0;r<e;){var a=t[r++];if((a.f&(Bt|yt))===0&&_r(a)&&(Wt=new Set,Ge(a),a.deps===null&&a.first===null&&a.nodes===null&&a.teardown===null&&a.ac===null&&ln(a),(Wt==null?void 0:Wt.size)>0)){Me.clear();for(const n of Wt){if((n.f&(Bt|yt))!==0)continue;const s=[n];let o=n.parent;for(;o!==null;)Wt.has(o)&&(Wt.delete(o),s.push(o)),o=o.parent;for(let u=s.length-1;u>=0;u--){const d=s[u];(d.f&(Bt|yt))===0&&Ge(d)}}Wt.clear()}}Wt=null}}function Va(t,e,r,a){if(!r.has(t)&&(r.add(t),t.reactions!==null))for(const n of t.reactions){const s=n.f;(s&_t)!==0?Va(n,e,r,a):(s&(or|re))!==0&&(s&ht)===0&&ra(n,e,a)&&(ot(n,ht),aa(n))}}function ra(t,e,r){const a=r.get(t);if(a!==void 0)return a;if(t.deps!==null)for(const n of t.deps){if(qe.call(e,n))return!0;if((n.f&_t)!==0&&ra(n,e,r))return r.set(n,!0),!0}return r.set(t,!1),!1}function aa(t){z.schedule(t)}function Wa(t,e){if(!((t.f&Gt)!==0&&(t.f&ft)!==0)){(t.f&ht)!==0?e.d.push(t):(t.f&ae)!==0&&e.m.push(t),ot(t,ft);for(var r=t.first;r!==null;)Wa(r,e),r=r.next}}function Ja(t){ot(t,ft);for(var e=t.first;e!==null;)Ja(e),e=e.next}function gs(t){let e=0,r=Le(0),a;return()=>{ia()&&(i(r),sn(()=>(e===0&&(a=xn(()=>t(()=>sr(r)))),e+=1,()=>{he(()=>{e-=1,e===0&&(a==null||a(),a=void 0,sr(r))})})))}}var xs=Ve|Xe;function ys(t,e,r,a){new bs(t,e,r,a)}var Dt,ta,Ot,Ie,wt,Lt,xt,Ct,le,Ne,ce,He,ur,cr,de,Nr,at,ms,ws,ks,Jr,kr,Er,Kr,Gr;class bs{constructor(e,r,a,n){K(this,at);Vt(this,"parent");Vt(this,"is_pending",!1);Vt(this,"transform_error");K(this,Dt);K(this,ta,null);K(this,Ot);K(this,Ie);K(this,wt);K(this,Lt,null);K(this,xt,null);K(this,Ct,null);K(this,le,null);K(this,Ne,0);K(this,ce,0);K(this,He,!1);K(this,ur,new Set);K(this,cr,new Set);K(this,de,null);K(this,Nr,gs(()=>(Y(this,de,Le(c(this,Ne))),()=>{Y(this,de,null)})));var s;Y(this,Dt,e),Y(this,Ot,r),Y(this,Ie,o=>{var u=X;u.b=this,u.f|=zr,a(o)}),this.parent=X.b,this.transform_error=n??((s=this.parent)==null?void 0:s.transform_error)??(o=>o),Y(this,wt,oa(()=>{et(this,at,Jr).call(this)},xs))}defer_effect(e){Ha(e,c(this,ur),c(this,cr))}is_rendered(){return!this.is_pending&&(!this.parent||this.parent.is_rendered())}has_pending_snippet(){return!!c(this,Ot).pending}update_pending_count(e,r){et(this,at,Kr).call(this,e,r),Y(this,Ne,c(this,Ne)+e),!(!c(this,de)||c(this,He))&&(Y(this,He,!0),he(()=>{Y(this,He,!1),c(this,de)&&Je(c(this,de),c(this,Ne))}))}get_effect_pending(){return c(this,Nr).call(this),i(c(this,de))}error(e){if(!c(this,Ot).onerror&&!c(this,Ot).failed)throw e;z!=null&&z.is_fork?(c(this,Lt)&&z.skip_effect(c(this,Lt)),c(this,xt)&&z.skip_effect(c(this,xt)),c(this,Ct)&&z.skip_effect(c(this,Ct)),z.on_fork_commit(()=>{et(this,at,Gr).call(this,e)})):et(this,at,Gr).call(this,e)}}Dt=new WeakMap,ta=new WeakMap,Ot=new WeakMap,Ie=new WeakMap,wt=new WeakMap,Lt=new WeakMap,xt=new WeakMap,Ct=new WeakMap,le=new WeakMap,Ne=new WeakMap,ce=new WeakMap,He=new WeakMap,ur=new WeakMap,cr=new WeakMap,de=new WeakMap,Nr=new WeakMap,at=new WeakSet,ms=function(){try{Y(this,Lt,Pt(()=>c(this,Ie).call(this,c(this,Dt))))}catch(e){this.error(e)}},ws=function(e){const r=c(this,Ot).failed;r&&Y(this,Ct,Pt(()=>{r(c(this,Dt),()=>e,()=>()=>{})}))},ks=function(){const e=c(this,Ot).pending;e&&(this.is_pending=!0,Y(this,xt,Pt(()=>e(c(this,Dt)))),he(()=>{var r=Y(this,le,document.createDocumentFragment()),a=_e();r.append(a),Y(this,Lt,et(this,at,Er).call(this,()=>Pt(()=>c(this,Ie).call(this,a)))),c(this,ce)===0&&(c(this,Dt).before(r),Y(this,le,null),$e(c(this,xt),()=>{Y(this,xt,null)}),et(this,at,kr).call(this,z))}))},Jr=function(){try{if(this.is_pending=this.has_pending_snippet(),Y(this,ce,0),Y(this,Ne,0),Y(this,Lt,Pt(()=>{c(this,Ie).call(this,c(this,Dt))})),c(this,ce)>0){var e=Y(this,le,document.createDocumentFragment());fa(c(this,Lt),e);const r=c(this,Ot).pending;Y(this,xt,Pt(()=>r(c(this,Dt))))}else et(this,at,kr).call(this,z)}catch(r){this.error(r)}},kr=function(e){this.is_pending=!1,e.transfer_effects(c(this,ur),c(this,cr))},Er=function(e){var r=X,a=Z,n=zt;ne(c(this,wt)),Ht(c(this,wt)),We(c(this,wt).ctx);try{return Oe.ensure(),e()}catch(s){return Ua(s),null}finally{ne(r),Ht(a),We(n)}},Kr=function(e,r){var a;if(!this.has_pending_snippet()){this.parent&&et(a=this.parent,at,Kr).call(a,e,r);return}Y(this,ce,c(this,ce)+e),c(this,ce)===0&&(et(this,at,kr).call(this,r),c(this,xt)&&$e(c(this,xt),()=>{Y(this,xt,null)}),c(this,le)&&(c(this,Dt).before(c(this,le)),Y(this,le,null)))},Gr=function(e){c(this,Lt)&&(Et(c(this,Lt)),Y(this,Lt,null)),c(this,xt)&&(Et(c(this,xt)),Y(this,xt,null)),c(this,Ct)&&(Et(c(this,Ct)),Y(this,Ct,null));var r=c(this,Ot).onerror;let a=c(this,Ot).failed;var n=!1,s=!1;const o=()=>{if(n){ls();return}n=!0,s&&is(),c(this,Ct)!==null&&$e(c(this,Ct),()=>{Y(this,Ct,null)}),et(this,at,Er).call(this,()=>{et(this,at,Jr).call(this)})},u=d=>{try{s=!0,r==null||r(d,o),s=!1}catch(f){ve(f,c(this,wt)&&c(this,wt).parent)}a&&Y(this,Ct,et(this,at,Er).call(this,()=>{try{return Pt(()=>{var f=X;f.b=this,f.f|=zr,a(c(this,Dt),()=>d,()=>o)})}catch(f){return ve(f,c(this,wt).parent),null}}))};he(()=>{var d;try{d=this.transform_error(e)}catch(f){ve(f,c(this,wt)&&c(this,wt).parent);return}d!==null&&typeof d=="object"&&typeof d.then=="function"?d.then(u,f=>ve(f,c(this,wt)&&c(this,wt).parent)):u(d)})};function Es(t,e,r,a){const n=na;var s=t.filter(y=>!y.settled);if(r.length===0&&s.length===0){a(e.map(n));return}var o=X,u=Ss(),d=s.length===1?s[0].promise:s.length>1?Promise.all(s.map(y=>y.promise)):null;function f(y){u();try{a(y)}catch(g){(o.f&Bt)===0&&ve(g,o)}Ar()}if(r.length===0){d.then(()=>f(e.map(n)));return}var _=Ka();function b(){Promise.all(r.map(y=>As(y))).then(y=>f([...e.map(n),...y])).catch(y=>ve(y,o)).finally(()=>_())}d?d.then(()=>{u(),b(),Ar()}):b()}function Ss(){var t=X,e=Z,r=zt,a=z;return function(s=!0){ne(t),Ht(e),We(r),s&&(t.f&Bt)===0&&(a==null||a.activate(),a==null||a.apply())}}function Ar(t=!0){ne(null),Ht(null),We(null),t&&(z==null||z.deactivate())}function Ka(){var t=X,e=t.b,r=z,a=e.is_rendered();return e.update_pending_count(1,r),r.increment(a,t),(n=!1)=>{e.update_pending_count(-1,r),r.decrement(a,t,n)}}function na(t){var e=_t|ht;return X!==null&&(X.f|=Xe),{ctx:zt,deps:null,effects:null,equals:Pa,f:e,fn:t,reactions:null,rv:0,v:ut,wv:0,parent:X,ac:null}}function As(t,e,r){let a=X;a===null&&Zn();var n=void 0,s=Le(ut),o=!Z,u=new Map;return Bs(()=>{var g;var d=X,f=Da();n=f.promise;try{Promise.resolve(t()).then(f.resolve,f.reject).finally(Ar)}catch(v){f.reject(v),Ar()}var _=z;if(o){if((d.f&Pe)!==0)var b=Ka();if(a.b.is_rendered())(g=u.get(_))==null||g.reject(se),u.delete(_);else{for(const v of u.values())v.reject(se);u.clear()}u.set(_,f)}const y=(v,w=void 0)=>{if(b){var h=w===se;b(h)}if(!(w===se||(d.f&Bt)!==0)){if(_.activate(),w)s.f|=pe,Je(s,w);else{(s.f&pe)!==0&&(s.f^=pe),Je(s,v);for(const[x,I]of u){if(u.delete(x),x===_)break;I.reject(se)}}_.deactivate()}};f.promise.then(y,v=>y(null,v||"unknown"))}),an(()=>{for(const d of u.values())d.reject(se)}),new Promise(d=>{function f(_){function b(){_===n?d(s):f(n)}_.then(b,b)}f(n)})}function Yt(t){const e=na(t);return un(e),e}function Ts(t){const e=na(t);return e.equals=Fa,e}function Is(t){var e=t.effects;if(e!==null){t.effects=null;for(var r=0;r<e.length;r+=1)Et(e[r])}}function sa(t){var e,r=X,a=t.parent;if(!xe&&a!==null&&(a.f&(Bt|yt))!==0)return os(),t.v;ne(a);try{t.f&=~De,Is(t),e=hn(t)}finally{ne(r)}return e}function Ga(t){var e=sa(t);if(!t.equals(e)&&(t.wv=vn(),(!(z!=null&&z.is_fork)||t.deps===null)&&(z!==null?z.capture(t,e,!0):t.v=e,t.deps===null))){ot(t,ft);return}xe||(vt!==null?(ia()||z!=null&&z.is_fork)&&vt.set(t,e):ea(t))}function Ns(t){var e,r;if(t.effects!==null)for(const a of t.effects)(a.teardown||a.ac)&&((e=a.teardown)==null||e.call(a),(r=a.ac)==null||r.abort(se),a.teardown=Ft,a.ac=null,lr(a,0),la(a))}function Ya(t){if(t.effects!==null)for(const e of t.effects)e.teardown&&Ge(e)}let Yr=new Set;const Me=new Map;let Za=!1;function Le(t,e){var r={f:0,v:t,reactions:null,equals:Pa,rv:0,wv:0};return r}function j(t,e){const r=Le(t);return un(r),r}function Cs(t,e=!1,r=!0){const a=Le(t);return e||(a.equals=Fa),a}function N(t,e,r=!1){Z!==null&&(!Kt||(Z.f&Hr)!==0)&&ja()&&(Z.f&(_t|re|or|Hr))!==0&&(Ut===null||!qe.call(Ut,t))&&ss();let a=r?lt(e):e;return Je(t,a,yr)}function Je(t,e,r=null){if(!t.equals(e)){Me.set(t,xe?e:t.v);var a=Oe.ensure();if(a.capture(t,e),(t.f&_t)!==0){const n=t;(t.f&ht)!==0&&sa(n),vt===null&&ea(n)}t.wv=vn(),Xa(t,ht,r),X!==null&&(X.f&ft)!==0&&(X.f&(Gt|ge))===0&&(Rt===null?Hs([t]):Rt.push(t)),!a.is_fork&&Yr.size>0&&!Za&&Ms()}return e}function Ms(){Za=!1;for(const t of Yr)(t.f&ft)!==0&&ot(t,ae),_r(t)&&Ge(t);Yr.clear()}function sr(t){N(t,t.v+1)}function Xa(t,e,r){var a=t.reactions;if(a!==null)for(var n=a.length,s=0;s<n;s++){var o=a[s],u=o.f,d=(u&ht)===0;if(d&&ot(o,e),(u&_t)!==0){var f=o;vt==null||vt.delete(f),(u&De)===0&&(u&jt&&(o.f|=De),Xa(f,ae,r))}else if(d){var _=o;(u&re)!==0&&Wt!==null&&Wt.add(_),r!==null?r.push(_):aa(_)}}}function lt(t){if(typeof t!="object"||t===null||Lr in t)return t;const e=Ra(t);if(e!==Vn&&e!==Wn)return t;var r=new Map,a=$a(t),n=j(0),s=Re,o=u=>{if(Re===s)return u();var d=Z,f=Re;Ht(null),wa(s);var _=u();return Ht(d),wa(f),_};return a&&r.set("length",j(t.length)),new Proxy(t,{defineProperty(u,d,f){(!("value"in f)||f.configurable===!1||f.enumerable===!1||f.writable===!1)&&as();var _=r.get(d);return _===void 0?o(()=>{var b=j(f.value);return r.set(d,b),b}):N(_,f.value,!0),!0},deleteProperty(u,d){var f=r.get(d);if(f===void 0){if(d in u){const _=o(()=>j(ut));r.set(d,_),sr(n)}}else N(f,ut),sr(n);return!0},get(u,d,f){var g;if(d===Lr)return t;var _=r.get(d),b=d in u;if(_===void 0&&(!b||(g=ar(u,d))!=null&&g.writable)&&(_=o(()=>{var v=lt(b?u[d]:ut),w=j(v);return w}),r.set(d,_)),_!==void 0){var y=i(_);return y===ut?void 0:y}return Reflect.get(u,d,f)},getOwnPropertyDescriptor(u,d){var f=Reflect.getOwnPropertyDescriptor(u,d);if(f&&"value"in f){var _=r.get(d);_&&(f.value=i(_))}else if(f===void 0){var b=r.get(d),y=b==null?void 0:b.v;if(b!==void 0&&y!==ut)return{enumerable:!0,configurable:!0,value:y,writable:!0}}return f},has(u,d){var y;if(d===Lr)return!0;var f=r.get(d),_=f!==void 0&&f.v!==ut||Reflect.has(u,d);if(f!==void 0||X!==null&&(!_||(y=ar(u,d))!=null&&y.writable)){f===void 0&&(f=o(()=>{var g=_?lt(u[d]):ut,v=j(g);return v}),r.set(d,f));var b=i(f);if(b===ut)return!1}return _},set(u,d,f,_){var $;var b=r.get(d),y=d in u;if(a&&d==="length")for(var g=f;g<b.v;g+=1){var v=r.get(g+"");v!==void 0?N(v,ut):g in u&&(v=o(()=>j(ut)),r.set(g+"",v))}if(b===void 0)(!y||($=ar(u,d))!=null&&$.writable)&&(b=o(()=>j(void 0)),N(b,lt(f)),r.set(d,b));else{y=b.v!==ut;var w=o(()=>lt(f));N(b,w)}var h=Reflect.getOwnPropertyDescriptor(u,d);if(h!=null&&h.set&&h.set.call(_,f),!y){if(a&&typeof d=="string"){var x=r.get("length"),I=Number(d);Number.isInteger(I)&&I>=x.v&&N(x,I+1)}sr(n)}return!0},ownKeys(u){i(n);var d=Reflect.ownKeys(u).filter(b=>{var y=r.get(b);return y===void 0||y.v!==ut});for(var[f,_]of r)_.v!==ut&&!(f in u)&&d.push(f);return d},setPrototypeOf(){ns()}})}var ya,Qa,tn,en;function $s(){if(ya===void 0){ya=window,Qa=/Firefox/.test(navigator.userAgent);var t=Element.prototype,e=Node.prototype,r=Text.prototype;tn=ar(e,"firstChild").get,en=ar(e,"nextSibling").get,ha(t)&&(t.__click=void 0,t.__className=void 0,t.__attributes=null,t.__style=void 0,t.__e=void 0),ha(r)&&(r.__t=void 0)}}function _e(t=""){return document.createTextNode(t)}function Ke(t){return tn.call(t)}function hr(t){return en.call(t)}function l(t,e){return Ke(t)}function Qe(t,e=!1){{var r=Ke(t);return r instanceof Comment&&r.data===""?hr(r):r}}function p(t,e=1,r=!1){let a=t;for(;e--;)a=hr(a);return a}function Rs(t){t.textContent=""}function rn(){return!1}function Ds(t,e,r){return document.createElementNS(Ma,t,void 0)}let ba=!1;function Os(){ba||(ba=!0,document.addEventListener("reset",t=>{Promise.resolve().then(()=>{var e;if(!t.defaultPrevented)for(const r of t.target.elements)(e=r.__on_r)==null||e.call(r)})},{capture:!0}))}function Rr(t){var e=Z,r=X;Ht(null),ne(null);try{return t()}finally{Ht(e),ne(r)}}function Ls(t,e,r,a=r){t.addEventListener(e,()=>Rr(r));const n=t.__on_r;n?t.__on_r=()=>{n(),a(!0)}:t.__on_r=()=>a(!0),Os()}function Ps(t){X===null&&(Z===null&&es(),ts()),xe&&Qn()}function Fs(t,e){var r=e.last;r===null?e.last=e.first=t:(r.next=t,t.prev=r,e.last=t)}function ye(t,e){var r=X;r!==null&&(r.f&yt)!==0&&(t|=yt);var a={ctx:zt,deps:null,nodes:null,f:t|ht|jt,first:null,fn:e,last:null,next:null,parent:r,b:r&&r.b,prev:null,teardown:null,wv:0,ac:null};z==null||z.register_created_effect(a);var n=a;if((t&ir)!==0)Fe!==null?Fe.push(a):Oe.ensure().schedule(a);else if(e!==null){try{Ge(a)}catch(o){throw Et(a),o}n.deps===null&&n.teardown===null&&n.nodes===null&&n.first===n.last&&(n.f&Xe)===0&&(n=n.first,(t&re)!==0&&(t&Ve)!==0&&n!==null&&(n.f|=Ve))}if(n!==null&&(n.parent=r,r!==null&&Fs(n,r),Z!==null&&(Z.f&_t)!==0&&(t&ge)===0)){var s=Z;(s.effects??(s.effects=[])).push(n)}return a}function ia(){return Z!==null&&!Kt}function an(t){const e=ye($r,null);return ot(e,ft),e.teardown=t,e}function Zt(t){Ps();var e=X.f,r=!Z&&(e&Gt)!==0&&(e&Pe)===0;if(r){var a=zt;(a.e??(a.e=[])).push(t)}else return nn(t)}function nn(t){return ye(ir|Gn,t)}function js(t){Oe.ensure();const e=ye(ge|Xe,t);return(r={})=>new Promise(a=>{r.outro?$e(e,()=>{Et(e),a(void 0)}):(Et(e),a(void 0))})}function Bs(t){return ye(or|Xe,t)}function sn(t,e=0){return ye($r|e,t)}function H(t,e=[],r=[],a=[]){Es(a,e,r,n=>{ye($r,()=>t(...n.map(i)))})}function oa(t,e=0){var r=ye(re|e,t);return r}function Pt(t){return ye(Gt|Xe,t)}function on(t){var e=t.teardown;if(e!==null){const r=xe,a=Z;ma(!0),Ht(null);try{e.call(null)}finally{ma(r),Ht(a)}}}function la(t,e=!1){var r=t.first;for(t.first=t.last=null;r!==null;){const n=r.ac;n!==null&&Rr(()=>{n.abort(se)});var a=r.next;(r.f&ge)!==0?r.parent=null:Et(r,e),r=a}}function Us(t){for(var e=t.first;e!==null;){var r=e.next;(e.f&Gt)===0&&Et(e),e=r}}function Et(t,e=!0){var r=!1;(e||(t.f&Kn)!==0)&&t.nodes!==null&&t.nodes.end!==null&&(zs(t.nodes.start,t.nodes.end),r=!0),ot(t,_a),la(t,e&&!r),lr(t,0);var a=t.nodes&&t.nodes.t;if(a!==null)for(const s of a)s.stop();on(t),t.f^=_a,t.f|=Bt;var n=t.parent;n!==null&&n.first!==null&&ln(t),t.next=t.prev=t.teardown=t.ctx=t.deps=t.fn=t.nodes=t.ac=t.b=null}function zs(t,e){for(;t!==null;){var r=t===e?null:hr(t);t.remove(),t=r}}function ln(t){var e=t.parent,r=t.prev,a=t.next;r!==null&&(r.next=a),a!==null&&(a.prev=r),e!==null&&(e.first===t&&(e.first=a),e.last===t&&(e.last=r))}function $e(t,e,r=!0){var a=[];dn(t,a,!0);var n=()=>{r&&Et(t),e&&e()},s=a.length;if(s>0){var o=()=>--s||n();for(var u of a)u.out(o)}else n()}function dn(t,e,r){if((t.f&yt)===0){t.f^=yt;var a=t.nodes&&t.nodes.t;if(a!==null)for(const u of a)(u.is_global||r)&&e.push(u);for(var n=t.first;n!==null;){var s=n.next;if((n.f&ge)===0){var o=(n.f&Ve)!==0||(n.f&Gt)!==0&&(t.f&re)!==0;dn(n,e,o?r:!1)}n=s}}}function da(t){fn(t,!0)}function fn(t,e){if((t.f&yt)!==0){t.f^=yt,(t.f&ft)===0&&(ot(t,ht),Oe.ensure().schedule(t));for(var r=t.first;r!==null;){var a=r.next,n=(r.f&Ve)!==0||(r.f&Gt)!==0;fn(r,n?e:!1),r=a}var s=t.nodes&&t.nodes.t;if(s!==null)for(const o of s)(o.is_global||e)&&o.in()}}function fa(t,e){if(t.nodes)for(var r=t.nodes.start,a=t.nodes.end;r!==null;){var n=r===a?null:hr(r);e.append(r),r=n}}let Sr=!1,xe=!1;function ma(t){xe=t}let Z=null,Kt=!1;function Ht(t){Z=t}let X=null;function ne(t){X=t}let Ut=null;function un(t){Z!==null&&(Ut===null?Ut=[t]:Ut.push(t))}let kt=null,It=0,Rt=null;function Hs(t){Rt=t}let cn=1,Ee=0,Re=Ee;function wa(t){Re=t}function vn(){return++cn}function _r(t){var e=t.f;if((e&ht)!==0)return!0;if(e&_t&&(t.f&=~De),(e&ae)!==0){for(var r=t.deps,a=r.length,n=0;n<a;n++){var s=r[n];if(_r(s)&&Ga(s),s.wv>t.wv)return!0}(e&jt)!==0&&vt===null&&ot(t,ft)}return!1}function pn(t,e,r=!0){var a=t.reactions;if(a!==null&&!(Ut!==null&&qe.call(Ut,t)))for(var n=0;n<a.length;n++){var s=a[n];(s.f&_t)!==0?pn(s,e,!1):e===s&&(r?ot(s,ht):(s.f&ft)!==0&&ot(s,ae),aa(s))}}function hn(t){var w;var e=kt,r=It,a=Rt,n=Z,s=Ut,o=zt,u=Kt,d=Re,f=t.f;kt=null,It=0,Rt=null,Z=(f&(Gt|ge))===0?t:null,Ut=null,We(t.ctx),Kt=!1,Re=++Ee,t.ac!==null&&(Rr(()=>{t.ac.abort(se)}),t.ac=null);try{t.f|=qr;var _=t.fn,b=_();t.f|=Pe;var y=t.deps,g=z==null?void 0:z.is_fork;if(kt!==null){var v;if(g||lr(t,It),y!==null&&It>0)for(y.length=It+kt.length,v=0;v<kt.length;v++)y[It+v]=kt[v];else t.deps=y=kt;if(ia()&&(t.f&jt)!==0)for(v=It;v<y.length;v++)((w=y[v]).reactions??(w.reactions=[])).push(t)}else!g&&y!==null&&It<y.length&&(lr(t,It),y.length=It);if(ja()&&Rt!==null&&!Kt&&y!==null&&(t.f&(_t|ae|ht))===0)for(v=0;v<Rt.length;v++)pn(Rt[v],t);if(n!==null&&n!==t){if(Ee++,n.deps!==null)for(let h=0;h<r;h+=1)n.deps[h].rv=Ee;if(e!==null)for(const h of e)h.rv=Ee;Rt!==null&&(a===null?a=Rt:a.push(...Rt))}return(t.f&pe)!==0&&(t.f^=pe),b}catch(h){return Ua(h)}finally{t.f^=qr,kt=e,It=r,Rt=a,Z=n,Ut=s,We(o),Kt=u,Re=d}}function qs(t,e){let r=e.reactions;if(r!==null){var a=zn.call(r,t);if(a!==-1){var n=r.length-1;n===0?r=e.reactions=null:(r[a]=r[n],r.pop())}}if(r===null&&(e.f&_t)!==0&&(kt===null||!qe.call(kt,e))){var s=e;(s.f&jt)!==0&&(s.f^=jt,s.f&=~De),s.v!==ut&&ea(s),Ns(s),lr(s,0)}}function lr(t,e){var r=t.deps;if(r!==null)for(var a=e;a<r.length;a++)qs(t,r[a])}function Ge(t){var e=t.f;if((e&Bt)===0){ot(t,ft);var r=X,a=Sr;X=t,Sr=!0;try{(e&(re|La))!==0?Us(t):la(t),on(t);var n=hn(t);t.teardown=typeof n=="function"?n:null,t.wv=cn;var s;Un&&fs&&(t.f&ht)!==0&&t.deps}finally{Sr=a,X=r}}}async function Vs(){await Promise.resolve(),hs()}function i(t){var e=t.f,r=(e&_t)!==0;if(Z!==null&&!Kt){var a=X!==null&&(X.f&Bt)!==0;if(!a&&(Ut===null||!qe.call(Ut,t))){var n=Z.deps;if((Z.f&qr)!==0)t.rv<Ee&&(t.rv=Ee,kt===null&&n!==null&&n[It]===t?It++:kt===null?kt=[t]:kt.push(t));else{(Z.deps??(Z.deps=[])).push(t);var s=t.reactions;s===null?t.reactions=[Z]:qe.call(s,Z)||s.push(Z)}}}if(xe&&Me.has(t))return Me.get(t);if(r){var o=t;if(xe){var u=o.v;return((o.f&ft)===0&&o.reactions!==null||gn(o))&&(u=sa(o)),Me.set(o,u),u}var d=(o.f&jt)===0&&!Kt&&Z!==null&&(Sr||(Z.f&jt)!==0),f=(o.f&Pe)===0;_r(o)&&(d&&(o.f|=jt),Ga(o)),d&&!f&&(Ya(o),_n(o))}if(vt!=null&&vt.has(t))return vt.get(t);if((t.f&pe)!==0)throw t.v;return t.v}function _n(t){if(t.f|=jt,t.deps!==null)for(const e of t.deps)(e.reactions??(e.reactions=[])).push(t),(e.f&_t)!==0&&(e.f&jt)===0&&(Ya(e),_n(e))}function gn(t){if(t.v===ut)return!0;if(t.deps===null)return!1;for(const e of t.deps)if(Me.has(e)||(e.f&_t)!==0&&gn(e))return!0;return!1}function xn(t){var e=Kt;try{return Kt=!0,t()}finally{Kt=e}}const Se=Symbol("events"),yn=new Set,Zr=new Set;function Ws(t,e,r,a={}){function n(s){if(a.capture||Xr.call(e,s),!s.cancelBubble)return Rr(()=>r==null?void 0:r.call(this,s))}return t.startsWith("pointer")||t.startsWith("touch")||t==="wheel"?he(()=>{e.addEventListener(t,n,a)}):e.addEventListener(t,n,a),n}function bn(t,e,r,a,n){var s={capture:a,passive:n},o=Ws(t,e,r,s);(e===document.body||e===window||e===document||e instanceof HTMLMediaElement)&&an(()=>{e.removeEventListener(t,o,s)})}function rt(t,e,r){(e[Se]??(e[Se]={}))[t]=r}function be(t){for(var e=0;e<t.length;e++)yn.add(t[e]);for(var r of Zr)r(t)}let ka=null;function Xr(t){var h,x;var e=this,r=e.ownerDocument,a=t.type,n=((h=t.composedPath)==null?void 0:h.call(t))||[],s=n[0]||t.target;ka=t;var o=0,u=ka===t&&t[Se];if(u){var d=n.indexOf(u);if(d!==-1&&(e===document||e===window)){t[Se]=e;return}var f=n.indexOf(e);if(f===-1)return;d<=f&&(o=d)}if(s=n[o]||t.target,s!==e){Hn(t,"currentTarget",{configurable:!0,get(){return s||r}});var _=Z,b=X;Ht(null),ne(null);try{for(var y,g=[];s!==null;){var v=s.assignedSlot||s.parentNode||s.host||null;try{var w=(x=s[Se])==null?void 0:x[a];w!=null&&(!s.disabled||t.target===s)&&w.call(s,t)}catch(I){y?g.push(I):y=I}if(t.cancelBubble||v===e||v===null)break;s=v}if(y){for(let I of g)queueMicrotask(()=>{throw I});throw y}}finally{t[Se]=e,delete t.currentTarget,Ht(_),ne(b)}}}var Na;const Fr=((Na=globalThis==null?void 0:globalThis.window)==null?void 0:Na.trustedTypes)&&globalThis.window.trustedTypes.createPolicy("svelte-trusted-html",{createHTML:t=>t});function Js(t){return(Fr==null?void 0:Fr.createHTML(t))??t}function mn(t){var e=Ds("template");return e.innerHTML=Js(t.replaceAll("<!>","<!---->")),e.content}function Tr(t,e){var r=X;r.nodes===null&&(r.nodes={start:t,end:e,a:null,t:null})}function S(t,e){var r=(e&jn)!==0,a=(e&Bn)!==0,n,s=!t.startsWith("<!>");return()=>{n===void 0&&(n=mn(s?t:"<!>"+t),r||(n=Ke(n)));var o=a||Qa?document.importNode(n,!0):n.cloneNode(!0);if(r){var u=Ke(o),d=o.lastChild;Tr(u,d)}else Tr(o,o);return o}}function Ks(t,e,r="svg"){var a=!t.startsWith("<!>"),n=`<${r}>${a?t:"<!>"+t}</${r}>`,s;return()=>{if(!s){var o=mn(n),u=Ke(o);s=Ke(u)}var d=s.cloneNode(!0);return Tr(d,d),d}}function wn(t,e){return Ks(t,e,"svg")}function kn(){var t=document.createDocumentFragment(),e=document.createComment(""),r=_e();return t.append(e,r),Tr(e,r),t}function E(t,e){t!==null&&t.before(e)}const Gs=["touchstart","touchmove"];function Ys(t){return Gs.includes(t)}function A(t,e){var r=e==null?"":typeof e=="object"?`${e}`:e;r!==(t.__t??(t.__t=t.nodeValue))&&(t.__t=r,t.nodeValue=`${r}`)}function Zs(t,e){return Xs(t,e)}const xr=new Map;function Xs(t,{target:e,anchor:r,props:a={},events:n,context:s,intro:o=!0,transformError:u}){$s();var d=void 0,f=js(()=>{var _=r??e.appendChild(_e());ys(_,{pending:()=>{}},g=>{At({});var v=zt;s&&(v.c=s),n&&(a.$$events=n),d=t(g,a)||{},Tt()},u);var b=new Set,y=g=>{for(var v=0;v<g.length;v++){var w=g[v];if(!b.has(w)){b.add(w);var h=Ys(w);for(const $ of[e,document]){var x=xr.get($);x===void 0&&(x=new Map,xr.set($,x));var I=x.get(w);I===void 0?($.addEventListener(w,Xr,{passive:h}),x.set(w,1)):x.set(w,I+1)}}}};return y(Mr(yn)),Zr.add(y),()=>{var h;for(var g of b)for(const x of[e,document]){var v=xr.get(x),w=v.get(g);--w==0?(x.removeEventListener(g,Xr),v.delete(g),v.size===0&&xr.delete(x)):v.set(g,w)}Zr.delete(y),_!==r&&((h=_.parentNode)==null||h.removeChild(_))}});return Qs.set(d,f),d}let Qs=new WeakMap;var Jt,te,Mt,Ce,vr,pr,Cr;class ti{constructor(e,r=!0){Vt(this,"anchor");K(this,Jt,new Map);K(this,te,new Map);K(this,Mt,new Map);K(this,Ce,new Set);K(this,vr,!0);K(this,pr,e=>{if(c(this,Jt).has(e)){var r=c(this,Jt).get(e),a=c(this,te).get(r);if(a)da(a),c(this,Ce).delete(r);else{var n=c(this,Mt).get(r);n&&(c(this,te).set(r,n.effect),c(this,Mt).delete(r),n.fragment.lastChild.remove(),this.anchor.before(n.fragment),a=n.effect)}for(const[s,o]of c(this,Jt)){if(c(this,Jt).delete(s),s===e)break;const u=c(this,Mt).get(o);u&&(Et(u.effect),c(this,Mt).delete(o))}for(const[s,o]of c(this,te)){if(s===r||c(this,Ce).has(s))continue;const u=()=>{if(Array.from(c(this,Jt).values()).includes(s)){var f=document.createDocumentFragment();fa(o,f),f.append(_e()),c(this,Mt).set(s,{effect:o,fragment:f})}else Et(o);c(this,Ce).delete(s),c(this,te).delete(s)};c(this,vr)||!a?(c(this,Ce).add(s),$e(o,u,!1)):u()}}});K(this,Cr,e=>{c(this,Jt).delete(e);const r=Array.from(c(this,Jt).values());for(const[a,n]of c(this,Mt))r.includes(a)||(Et(n.effect),c(this,Mt).delete(a))});this.anchor=e,Y(this,vr,r)}ensure(e,r){var a=z,n=rn();if(r&&!c(this,te).has(e)&&!c(this,Mt).has(e))if(n){var s=document.createDocumentFragment(),o=_e();s.append(o),c(this,Mt).set(e,{effect:Pt(()=>r(o)),fragment:s})}else c(this,te).set(e,Pt(()=>r(this.anchor)));if(c(this,Jt).set(a,e),n){for(const[u,d]of c(this,te))u===e?a.unskip_effect(d):a.skip_effect(d);for(const[u,d]of c(this,Mt))u===e?a.unskip_effect(d.effect):a.skip_effect(d.effect);a.oncommit(c(this,pr)),a.ondiscard(c(this,Cr))}else c(this,pr).call(this,a)}}Jt=new WeakMap,te=new WeakMap,Mt=new WeakMap,Ce=new WeakMap,vr=new WeakMap,pr=new WeakMap,Cr=new WeakMap;function G(t,e,r=!1){var a=new ti(t),n=r?Ve:0;function s(o,u){a.ensure(o,u)}oa(()=>{var o=!1;e((u,d=0)=>{o=!0,s(d,u)}),o||s(-1,null)},n)}function bt(t,e){return e}function ei(t,e,r){for(var a=[],n=e.length,s,o=e.length,u=0;u<n;u++){let b=e[u];$e(b,()=>{if(s){if(s.pending.delete(b),s.done.add(b),s.pending.size===0){var y=t.outrogroups;Qr(t,Mr(s.done)),y.delete(s),y.size===0&&(t.outrogroups=null)}}else o-=1},!1)}if(o===0){var d=a.length===0&&r!==null;if(d){var f=r,_=f.parentNode;Rs(_),_.append(f),t.items.clear()}Qr(t,e,!d)}else s={pending:new Set(e),done:new Set},(t.outrogroups??(t.outrogroups=new Set)).add(s)}function Qr(t,e,r=!0){var a;if(t.pending.size>0){a=new Set;for(const o of t.pending.values())for(const u of o)a.add(t.items.get(u).e)}for(var n=0;n<e.length;n++){var s=e[n];if(a!=null&&a.has(s)){s.f|=ee;const o=document.createDocumentFragment();fa(s,o)}else Et(e[n],r)}}var Ea;function mt(t,e,r,a,n,s=null){var o=t,u=new Map,d=(e&Ca)!==0;if(d){var f=t;o=f.appendChild(_e())}var _=null,b=Ts(()=>{var $=r();return $a($)?$:$==null?[]:Mr($)}),y,g=new Map,v=!0;function w($){(I.effect.f&Bt)===0&&(I.pending.delete($),I.fallback=_,ri(I,y,o,e,a),_!==null&&(y.length===0?(_.f&ee)===0?da(_):(_.f^=ee,rr(_,null,o)):$e(_,()=>{_=null})))}function h($){I.pending.delete($)}var x=oa(()=>{y=i(b);for(var $=y.length,C=new Set,O=z,P=rn(),R=0;R<$;R+=1){var q=y[R],m=a(q,R),T=v?null:u.get(m);T?(T.v&&Je(T.v,q),T.i&&Je(T.i,R),P&&O.unskip_effect(T.e)):(T=ai(u,v?o:Ea??(Ea=_e()),q,m,R,n,e,r),v||(T.e.f|=ee),u.set(m,T)),C.add(m)}if($===0&&s&&!_&&(v?_=Pt(()=>s(o)):(_=Pt(()=>s(Ea??(Ea=_e()))),_.f|=ee)),$>C.size&&Xn(),!v)if(g.set(O,C),P){for(const[k,M]of u)C.has(k)||O.skip_effect(M.e);O.oncommit(w),O.ondiscard(h)}else w(O);i(b)}),I={effect:x,items:u,pending:g,outrogroups:null,fallback:_};v=!1}function tr(t){for(;t!==null&&(t.f&Gt)===0;)t=t.next;return t}function ri(t,e,r,a,n){var T,k,M,B,D,L,F,U,J;var s=(a&Pn)!==0,o=e.length,u=t.items,d=tr(t.effect.first),f,_=null,b,y=[],g=[],v,w,h,x;if(s)for(x=0;x<o;x+=1)v=e[x],w=n(v,x),h=u.get(w).e,(h.f&ee)===0&&((k=(T=h.nodes)==null?void 0:T.a)==null||k.measure(),(b??(b=new Set)).add(h));for(x=0;x<o;x+=1){if(v=e[x],w=n(v,x),h=u.get(w).e,t.outrogroups!==null)for(const V of t.outrogroups)V.pending.delete(h),V.done.delete(h);if((h.f&yt)!==0&&(da(h),s&&((B=(M=h.nodes)==null?void 0:M.a)==null||B.unfix(),(b??(b=new Set)).delete(h))),(h.f&ee)!==0)if(h.f^=ee,h===d)rr(h,null,r);else{var I=_?_.next:d;h===t.effect.last&&(t.effect.last=h.prev),h.prev&&(h.prev.next=h.next),h.next&&(h.next.prev=h.prev),fe(t,_,h),fe(t,h,I),rr(h,I,r),_=h,y=[],g=[],d=tr(_.next);continue}if(h!==d){if(f!==void 0&&f.has(h)){if(y.length<g.length){var $=g[0],C;_=$.prev;var O=y[0],P=y[y.length-1];for(C=0;C<y.length;C+=1)rr(y[C],$,r);for(C=0;C<g.length;C+=1)f.delete(g[C]);fe(t,O.prev,P.next),fe(t,_,O),fe(t,P,$),d=$,_=P,x-=1,y=[],g=[]}else f.delete(h),rr(h,d,r),fe(t,h.prev,h.next),fe(t,h,_===null?t.effect.first:_.next),fe(t,_,h),_=h;continue}for(y=[],g=[];d!==null&&d!==h;)(f??(f=new Set)).add(d),g.push(d),d=tr(d.next);if(d===null)continue}(h.f&ee)===0&&y.push(h),_=h,d=tr(h.next)}if(t.outrogroups!==null){for(const V of t.outrogroups)V.pending.size===0&&(Qr(t,Mr(V.done)),(D=t.outrogroups)==null||D.delete(V));t.outrogroups.size===0&&(t.outrogroups=null)}if(d!==null||f!==void 0){var R=[];if(f!==void 0)for(h of f)(h.f&yt)===0&&R.push(h);for(;d!==null;)(d.f&yt)===0&&d!==t.fallback&&R.push(d),d=tr(d.next);var q=R.length;if(q>0){var m=(a&Ca)!==0&&o===0?r:null;if(s){for(x=0;x<q;x+=1)(F=(L=R[x].nodes)==null?void 0:L.a)==null||F.measure();for(x=0;x<q;x+=1)(J=(U=R[x].nodes)==null?void 0:U.a)==null||J.fix()}ei(t,R,m)}}s&&he(()=>{var V,W;if(b!==void 0)for(h of b)(W=(V=h.nodes)==null?void 0:V.a)==null||W.apply()})}function ai(t,e,r,a,n,s,o,u){var d=(o&On)!==0?(o&Fn)===0?Cs(r,!1,!1):Le(r):null,f=(o&Ln)!==0?Le(n):null;return{v:d,i:f,e:Pt(()=>(s(e,d??r,f??n,u),()=>{t.delete(a)}))}}function rr(t,e,r){if(t.nodes)for(var a=t.nodes.start,n=t.nodes.end,s=e&&(e.f&ee)===0?e.nodes.start:r;a!==null;){var o=hr(a);if(s.before(a),a===n)return;a=o}}function fe(t,e,r){e===null?t.effect.first=r:e.next=r,r===null?t.effect.last=e:r.prev=e}function ni(t,e,r){var a=t==null?"":""+t;return a===""?null:a}function pt(t,e,r,a,n,s){var o=t.__className;if(o!==r||o===void 0){var u=ni(r);u==null?t.removeAttribute("class"):t.className=u,t.__className=r}return s}const si=Symbol("is custom element"),ii=Symbol("is html"),oi=Yn?"progress":"PROGRESS";function jr(t,e){var r=ua(t);r.value===(r.value=e??void 0)||t.value===e&&(e!==0||t.nodeName!==oi)||(t.value=e??"")}function li(t,e){var r=ua(t);r.checked!==(r.checked=e??void 0)&&(t.checked=e)}function di(t,e,r,a){var n=ua(t);n[e]!==(n[e]=r)&&(r==null?t.removeAttribute(e):typeof r!="string"&&fi(t).includes(e)?t[e]=r:t.setAttribute(e,r))}function ua(t){return t.__attributes??(t.__attributes={[si]:t.nodeName.includes("-"),[ii]:t.namespaceURI===Ma})}var Sa=new Map;function fi(t){var e=t.getAttribute("is")||t.nodeName,r=Sa.get(e);if(r)return r;Sa.set(e,r=[]);for(var a,n=t,s=Element.prototype;s!==n;){a=qn(n);for(var o in a)a[o].set&&r.push(o);n=Ra(n)}return r}function Ye(t,e,r=e){var a=new WeakSet;Ls(t,"input",async n=>{var s=n?t.defaultValue:t.value;if(s=Br(t)?Ur(s):s,r(s),z!==null&&a.add(z),await Vs(),s!==(s=e())){var o=t.selectionStart,u=t.selectionEnd,d=t.value.length;if(t.value=s??"",u!==null){var f=t.value.length;o===u&&u===d&&f>d?(t.selectionStart=f,t.selectionEnd=f):(t.selectionStart=o,t.selectionEnd=Math.min(u,f))}}}),xn(e)==null&&t.value&&(r(Br(t)?Ur(t.value):t.value),z!==null&&a.add(z)),sn(()=>{var n=e();if(t===document.activeElement){var s=z;if(a.has(s))return}Br(t)&&n===Ur(t.value)||t.type==="date"&&!n&&!t.value||n!==t.value&&(t.value=n??"")})}function Br(t){var e=t.type;return e==="number"||e==="range"}function Ur(t){return t===""?null:+t}function ui(t,e,r,a){var n=a,s=!0,o=()=>(s&&(s=!1,n=a),n),u;u=t[e],u===void 0&&a!==void 0&&(u=o());var d;return d=()=>{var f=t[e];return f===void 0?o():(s=!0,f)},d}const ci="/aonx/v1";function En(){return localStorage.getItem("kagami_session_token")||""}async function ca(t,e,r=null,a={}){const n=a.token||En(),s={"Content-Type":"application/json"};n&&(s.Authorization=`Bearer ${n}`);const o=await fetch(ci+e,{method:t,headers:s,body:r?JSON.stringify(r):null});if(a.raw)return o;let u=null;const d=await o.text();if(d)try{u=JSON.parse(d)}catch{u=d}return{status:o.status,data:u}}const St=(t,e)=>ca("GET",t,null,e),Ze=(t,e,r)=>ca("POST",t,e,r),Aa=(t,e,r)=>ca("PATCH",t,e,r);async function vi(){const t=await fetch("/metrics");if(!t.ok)return new Map;const e=await t.text(),r=new Map;for(const a of e.split(`
+`)){if(!a||a.startsWith("#"))continue;const n=a.match(/^([a-zA-Z_:][a-zA-Z0-9_:]*(?:\{[^}]*\})?)\s+([\d.eE+-]+(?:nan|inf)?)/i);n&&r.set(n[1],parseFloat(n[2]))}return r}const it=lt({authToken:localStorage.getItem("kagami_auth_token")||"",sessionToken:localStorage.getItem("kagami_session_token")||"",username:localStorage.getItem("kagami_username")||"",acl:localStorage.getItem("kagami_acl")||"",loggedIn:!!localStorage.getItem("kagami_session_token")});function Sn(){localStorage.setItem("kagami_auth_token",it.authToken),localStorage.setItem("kagami_session_token",it.sessionToken),localStorage.setItem("kagami_username",it.username),localStorage.setItem("kagami_acl",it.acl)}async function pi(t,e){var s,o;const r=await Ze("/auth/login",{username:t,password:e});if(r.status!==201)return{ok:!1,error:((s=r.data)==null?void 0:s.reason)||"Login failed"};const a=r.data.token,n=await Ze("/session",{client_name:"kagami-admin",client_version:"1.0",hdid:"admin-dashboard-"+crypto.randomUUID().slice(0,8),auth:a});return n.status!==201?{ok:!1,error:((o=n.data)==null?void 0:o.reason)||"Session creation failed"}:(it.authToken=a,it.sessionToken=n.data.token,it.username=r.data.username,it.acl=r.data.acl,it.loggedIn=!0,Sn(),{ok:!0})}async function hi(){it.authToken&&await Ze("/auth/logout",null,{token:it.authToken}),it.authToken="",it.sessionToken="",it.username="",it.acl="",it.loggedIn=!1,Sn()}var _i=wn('<path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"></path>'),gi=wn('<path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M4 6h16M4 12h16M4 18h16"></path>'),xi=S('<button><span class="text-base"> </span> </button>'),yi=S('<div class="fixed inset-0 bg-black/50 z-30 md:hidden"></div>'),bi=S('<button class="md:hidden fixed top-3 left-3 z-50 p-2 bg-gray-800 rounded-lg border border-gray-700" aria-label="Toggle navigation"><svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24"><!></svg></button>  <aside><div class="p-4 border-b border-gray-800"><h1 class="text-lg font-bold tracking-tight">Kagami</h1> <p class="text-xs text-gray-500"> </p></div> <nav class="flex-1 overflow-y-auto py-2"><!> <a href="/grafana/" target="_blank" rel="noopener" class="w-full flex items-center gap-3 px-4 py-2 text-sm text-gray-400 hover:bg-gray-800/50 hover:text-gray-200"><span class="text-base"></span> Monitoring <span class="text-xs text-gray-600 ml-auto"></span></a></nav> <div class="p-3 border-t border-gray-800"><button class="w-full px-3 py-1.5 text-xs text-gray-400 hover:text-red-400 hover:bg-gray-800 rounded transition-colors">Logout</button></div></aside> <!>',1);function mi(t,e){At(e,!0);let r=ui(e,"currentPage",3,"dashboard"),a=j(!1);const n=[{page:"dashboard",label:"Dashboard",icon:"🏠"},{page:"sessions",label:"Sessions",icon:"👥"},{page:"traffic",label:"Traffic",icon:"💬"},{page:"config",label:"Config",icon:"⚙"},{page:"bans",label:"Bans",icon:"🚫"},{page:"moderation",label:"Moderation",icon:"🛡"},{page:"areas",label:"Areas",icon:"🗺"},{page:"users",label:"Users",icon:"🔑"},{page:"firewall",label:"Firewall",icon:"🔥"},{page:"content",label:"Content",icon:"📁"}];function s(T){window.location.hash="#/"+T,N(a,!1)}async function o(){await hi(),window.location.hash="#/login"}var u=bi(),d=Qe(u),f=l(d),_=l(f);{var b=T=>{var k=_i();E(T,k)},y=T=>{var k=gi();E(T,k)};G(_,T=>{i(a)?T(b):T(y,-1)})}var g=p(d,2),v=l(g),w=p(l(v),2),h=l(w),x=p(v,2),I=l(x);mt(I,17,()=>n,bt,(T,k)=>{let M=()=>i(k).page,B=()=>i(k).label,D=()=>i(k).icon;var L=xi(),F=l(L),U=l(F),J=p(F);H(()=>{pt(L,1,`w-full flex items-center gap-3 px-4 py-2 text-sm transition-colors
+               ${r()===M()?"bg-gray-800 text-white":"text-gray-400 hover:bg-gray-800/50 hover:text-gray-200"}`),A(U,D()),A(J,` ${B()??""}`)}),rt("click",L,()=>s(M())),E(T,L)});var $=p(I,2),C=l($);C.textContent="📊";var O=p(C,2);O.textContent="↗";var P=p(x,2),R=l(P),q=p(g,2);{var m=T=>{var k=yi();rt("click",k,()=>N(a,!1)),E(T,k)};G(q,T=>{i(a)&&T(m)})}H(()=>{pt(g,1,`fixed md:static inset-y-0 left-0 z-40 w-56 bg-gray-900 border-r border-gray-800
+         flex flex-col transition-transform duration-200 ease-in-out
+         ${i(a)?"translate-x-0":"-translate-x-full"} md:translate-x-0`),A(h,`${it.username??""} · ${it.acl??""}`)}),rt("click",d,()=>N(a,!i(a))),rt("click",R,o),E(t,u),Tt()}be(["click"]);var wi=S('<div class="text-sm text-red-400 bg-red-400/10 rounded-lg px-3 py-2"> </div>'),ki=S(`<div class="min-h-screen bg-gray-950 flex items-center justify-center px-4"><div class="w-full max-w-sm"><div class="bg-gray-900 rounded-xl border border-gray-800 p-6 shadow-2xl"><h1 class="text-xl font-bold text-white mb-1">Kagami</h1> <p class="text-sm text-gray-500 mb-6">Server Administration</p> <form class="space-y-4"><div><label for="username" class="block text-xs font-medium text-gray-400 mb-1">Username</label> <input id="username" type="text" required="" autocomplete="username" class="w-full px-3 py-2 bg-gray-800 border border-gray-700 rounded-lg text-sm text-white
+                   placeholder-gray-500 focus:outline-none focus:border-blue-500 focus:ring-1 focus:ring-blue-500" placeholder="root"/></div> <div><label for="password" class="block text-xs font-medium text-gray-400 mb-1">Password</label> <input id="password" type="password" required="" autocomplete="current-password" class="w-full px-3 py-2 bg-gray-800 border border-gray-700 rounded-lg text-sm text-white
+                   placeholder-gray-500 focus:outline-none focus:border-blue-500 focus:ring-1 focus:ring-blue-500"/></div> <!> <button type="submit" class="w-full py-2 px-4 bg-blue-600 hover:bg-blue-500 disabled:bg-gray-700 disabled:text-gray-500
+                 text-white text-sm font-medium rounded-lg transition-colors"> </button></form></div></div></div>`);function Ei(t,e){At(e,!0);let r=j(""),a=j(""),n=j(""),s=j(!1);async function o($){$.preventDefault(),N(n,""),N(s,!0);const C=await pi(i(r),i(a));N(s,!1),C.ok||N(n,C.error,!0)}var u=ki(),d=l(u),f=l(d),_=p(l(f),4),b=l(_),y=p(l(b),2),g=p(b,2),v=p(l(g),2),w=p(g,2);{var h=$=>{var C=wi(),O=l(C);H(()=>A(O,i(n))),E($,C)};G(w,$=>{i(n)&&$(h)})}var x=p(w,2),I=l(x);H(()=>{x.disabled=i(s),A(I,i(s)?"Signing in...":"Sign in")}),bn("submit",_,o),Ye(y,()=>i(r),$=>N(r,$)),Ye(v,()=>i(a),$=>N(a,$)),E(t,u),Tt()}const er=(t,e=Ft,r=Ft,a=Ft)=>{var n=Si(),s=l(n),o=l(s),u=p(o),d=l(u),f=p(s,2),_=l(f);H(()=>{A(o,r()),A(d,a()),A(_,e())}),E(t,n)};var Si=S('<div class="bg-gray-900 rounded-xl border border-gray-800 p-4"><div class="text-2xl font-bold tabular-nums"> <span class="text-sm text-gray-500 font-normal"> </span></div> <div class="text-xs text-gray-500 mt-1"> </div></div>'),Ai=S('<p class="text-gray-500">Loading...</p>'),Ti=S('<span class="text-sm text-gray-400"> </span>'),Ii=S('<div class="bg-gray-900 rounded-xl border border-gray-800 p-4"><div class="flex flex-wrap items-baseline gap-x-4 gap-y-1"><span class="text-lg font-semibold"> </span> <span class="text-xs text-gray-500"> </span> <!></div></div>'),Ni=S('<span class="ml-2 text-xs px-1.5 py-0.5 rounded bg-blue-900/50 text-blue-300"> </span>'),Ci=S('<span class="ml-1 text-xs px-1.5 py-0.5 rounded bg-yellow-900/50 text-yellow-300"> </span>'),Mi=S('<div class="px-4 py-2 flex items-center justify-between text-sm"><div><span class="font-medium"> </span> <!> <!></div> <span class="text-gray-500"> </span></div>'),$i=S('<tr class="hover:bg-gray-800/30"><td class="px-4 py-1.5"> </td><td class="px-4 py-1.5 text-gray-400"> </td><td class="px-4 py-1.5 text-gray-400"> </td><td class="px-4 py-1.5 text-right text-gray-500"> </td></tr>'),Ri=S('<div class="px-4 py-2 text-xs text-gray-500 border-t border-gray-800"> <a href="#/sessions" class="text-blue-400 hover:underline">View all</a></div>'),Di=S('<div class="bg-gray-900 rounded-xl border border-gray-800 overflow-hidden"><div class="px-4 py-3 border-b border-gray-800"><h3 class="text-sm font-semibold text-gray-300">Active Sessions</h3></div> <div class="overflow-x-auto"><table class="w-full text-sm"><thead><tr class="text-left text-xs text-gray-500 border-b border-gray-800"><th class="px-4 py-2">Name</th><th class="px-4 py-2">Protocol</th><th class="px-4 py-2">Area</th><th class="px-4 py-2 text-right">Idle</th></tr></thead><tbody class="divide-y divide-gray-800/50"></tbody></table></div> <!></div>'),Oi=S('<!> <div class="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-5 gap-3"><!> <!> <!> <!> <!></div> <div class="bg-gray-900 rounded-xl border border-gray-800 overflow-hidden"><div class="px-4 py-3 border-b border-gray-800"><h3 class="text-sm font-semibold text-gray-300">Areas</h3></div> <div class="divide-y divide-gray-800"></div></div> <!>',1),Li=S('<div class="space-y-6"><h2 class="text-2xl font-bold">Dashboard</h2> <!></div>');function Pi(t,e){At(e,!0);let r=j(null),a=j(lt([])),n=j(lt([])),s=j(lt(new Map)),o=j(!0);async function u(){var x;const[g,v,w,h]=await Promise.all([St("/server"),St("/admin/sessions"),St("/areas"),vi()]);g.status===200&&N(r,g.data,!0),v.status===200&&N(a,v.data,!0),w.status===200&&N(n,((x=w.data)==null?void 0:x.areas)||[],!0),N(s,h,!0),N(o,!1)}Zt(()=>{u();const g=setInterval(u,1e4);return()=>clearInterval(g)});function d(g){return i(s).get(g)??0}var f=Li(),_=p(l(f),2);{var b=g=>{var v=Ai();E(g,v)},y=g=>{var v=Oi(),w=Qe(v);{var h=k=>{var M=Ii(),B=l(M),D=l(B),L=l(D),F=p(D,2),U=l(F),J=p(F,2);{var V=W=>{var tt=Ti(),nt=l(tt);H(()=>A(nt,`— ${i(r).description??""}`)),E(W,tt)};G(J,W=>{i(r).description&&W(V)})}H(()=>{A(L,i(r).name),A(U,`v${i(r).version??""}`)}),E(k,M)};G(w,k=>{i(r)&&k(h)})}var x=p(w,2),I=l(x);er(I,()=>"Players",()=>{var k;return((k=i(r))==null?void 0:k.online)??0},()=>{var k;return"/"+(((k=i(r))==null?void 0:k.max)??"?")});var $=p(I,2);{let k=Yt(()=>d("kagami_ws_connections"));er($,()=>"AO2 Clients",()=>i(k),()=>"")}var C=p($,2);er(C,()=>"REST Sessions",()=>i(a).length,()=>"");var O=p(C,2);er(O,()=>"Areas",()=>i(n).length,()=>"");var P=p(O,2);{let k=Yt(()=>d("kagami_sessions_moderators"));er(P,()=>"Moderators",()=>i(k),()=>"")}var R=p(x,2),q=p(l(R),2);mt(q,21,()=>i(n),bt,(k,M)=>{var B=Mi(),D=l(B),L=l(D),F=l(L),U=p(L,2);{var J=Q=>{var st=Ni(),ct=l(st);H(()=>A(ct,i(M).status)),E(Q,st)};G(U,Q=>{i(M).status&&i(M).status!=="IDLE"&&Q(J)})}var V=p(U,2);{var W=Q=>{var st=Ci(),ct=l(st);H(()=>A(ct,i(M).locked)),E(Q,st)};G(V,Q=>{i(M).locked&&i(M).locked!=="FREE"&&Q(W)})}var tt=p(D,2),nt=l(tt);H(()=>{A(F,i(M).name),A(nt,`${i(M).players??0??""} players`)}),E(k,B)});var m=p(R,2);{var T=k=>{var M=Di(),B=p(l(M),2),D=l(B),L=p(l(D));mt(L,21,()=>i(a).slice(0,20),bt,(J,V)=>{var W=$i(),tt=l(W),nt=l(tt),Q=p(tt),st=l(Q),ct=p(Q),qt=l(ct),me=p(ct),gt=l(me);H(()=>{A(nt,i(V).display_name||"(anonymous)"),A(st,i(V).protocol),A(qt,i(V).area),A(gt,`${i(V).idle_seconds??""}s`)}),E(J,W)});var F=p(B,2);{var U=J=>{var V=Ri(),W=l(V);H(()=>A(W,`Showing 20 of ${i(a).length??""} — `)),E(J,V)};G(F,J=>{i(a).length>20&&J(U)})}E(k,M)};G(m,k=>{i(a).length>0&&k(T)})}E(g,v)};G(_,g=>{i(o)?g(b):g(y,-1)})}E(t,f),Tt()}const An=(t,e=Ft,r=Ft,a=Ft)=>{var n=ji();mt(n,21,()=>Object.entries(a()),bt,(s,o)=>{var u=Yt(()=>Oa(i(o),2));let d=()=>i(u)[0],f=()=>i(u)[1];var _=kn(),b=Qe(_);{var y=w=>{var h=Fi(),x=l(h),I=l(x),$=p(x,2),C=l($);An(C,a,d,f),H(()=>A(I,d())),E(w,h)},g=Yt(()=>typeof f()=="object"&&f()!==null&&!Array.isArray(f())),v=w=>{Tn(w,a,d,f)};G(b,w=>{i(g)?w(y):w(v,-1)})}E(s,_)}),E(t,n)},Tn=(t,e=Ft,r=Ft,a=Ft)=>{var n=qi(),s=l(n),o=l(s),u=p(s,2);{var d=g=>{var v=Bi();H(()=>li(v,a())),rt("change",v,w=>e()[r()]=w.target.checked),E(g,v)},f=g=>{var v=Ui();H(()=>jr(v,a())),rt("change",v,w=>e()[r()]=Number(w.target.value)),E(g,v)},_=g=>{var v=zi();H(w=>jr(v,w),[()=>JSON.stringify(a(),null,2)]),rt("change",v,w=>{try{e()[r()]=JSON.parse(w.target.value)}catch{}}),E(g,v)},b=Yt(()=>Array.isArray(a())),y=g=>{var v=Hi();H(()=>jr(v,a()??"")),rt("input",v,w=>e()[r()]=w.target.value),E(g,v)};G(u,g=>{typeof a()=="boolean"?g(d):typeof a()=="number"?g(f,1):i(b)?g(_,2):g(y,-1)})}H(()=>{di(s,"title",r()),A(o,r())}),E(t,n)};var Fi=S('<details class="ml-2 border-l border-gray-800 pl-3"><summary class="text-xs text-gray-400 cursor-pointer hover:text-gray-200"> </summary> <div class="mt-2"><!></div></details>'),ji=S('<div class="space-y-2"></div>'),Bi=S('<input type="checkbox" class="rounded bg-gray-800 border-gray-600"/>'),Ui=S('<input type="number" step="any" class="flex-1 px-2 py-1 text-xs bg-gray-800 border border-gray-700 rounded text-white"/>'),zi=S('<textarea class="flex-1 px-2 py-1 text-xs bg-gray-800 border border-gray-700 rounded text-white font-mono" rows="2"></textarea>'),Hi=S('<input type="text" class="flex-1 px-2 py-1 text-xs bg-gray-800 border border-gray-700 rounded text-white"/>'),qi=S('<div class="flex items-center gap-2 py-1"><label class="text-xs text-gray-400 w-48 shrink-0 truncate"> </label> <!></div>'),Vi=S('<div class="text-sm text-red-400 bg-red-400/10 rounded-lg px-3 py-2"> </div>'),Wi=S('<pre class="text-xs text-green-400 bg-green-400/10 rounded-lg px-3 py-2 whitespace-pre-wrap"> </pre>'),Ji=S('<p class="text-gray-500">Loading configuration...</p>'),Ki=S('<div class="bg-gray-900 rounded-xl border border-gray-800 overflow-hidden"><details><summary class="px-4 py-3 cursor-pointer hover:bg-gray-800/50 text-sm font-medium flex items-center justify-between"><span> </span> <span class="text-xs text-gray-500"> </span></summary> <div class="px-4 py-3 border-t border-gray-800 bg-gray-950/50"><!></div></details></div>'),Gi=S('<div class="space-y-3"></div>'),Yi=S('<div class="space-y-4"><div class="flex items-center justify-between flex-wrap gap-2"><h2 class="text-2xl font-bold">Configuration</h2> <div class="flex gap-2"><button class="px-3 py-1.5 text-sm bg-gray-700 hover:bg-gray-600 disabled:opacity-50 rounded-lg transition-colors">Reload Server</button> <button class="px-3 py-1.5 text-sm bg-blue-600 hover:bg-blue-500 disabled:opacity-50 rounded-lg transition-colors"> </button></div></div> <!> <!> <!></div>');function Zi(t,e){At(e,!0);let r=j(null),a=j(null),n=j(!0),s=j(!1),o=j(""),u=j("");async function d(){var T;N(n,!0);const m=await St("/admin/config");m.status===200?(N(r,m.data,!0),N(a,JSON.parse(JSON.stringify(m.data)),!0)):N(u,((T=m.data)==null?void 0:T.reason)||"Failed to load config",!0),N(n,!1)}Zt(()=>{d()});function f(m,T){const k={};for(const M of Object.keys(T))if(typeof T[M]=="object"&&T[M]!==null&&!Array.isArray(T[M])&&typeof m[M]=="object"&&m[M]!==null){const B=f(m[M],T[M]);Object.keys(B).length>0&&(k[M]=B)}else JSON.stringify(T[M])!==JSON.stringify(m[M])&&(k[M]=T[M]);return k}async function _(){var k,M;N(u,""),N(o,"");const m=f(i(a),i(r));if(Object.keys(m).length===0){N(o,"No changes to save.");return}N(s,!0);const T=await Aa("/admin/config",m);N(s,!1),T.status===200?(N(o,((k=T.data)==null?void 0:k.reload_summary)||"Config saved and reloaded.",!0),N(a,JSON.parse(JSON.stringify(i(r))),!0)):N(u,((M=T.data)==null?void 0:M.reason)||"Failed to save config",!0)}async function b(){var T,k;N(o,""),N(s,!0);const m=await Aa("/admin/config",{});N(s,!1),m.status===200?N(o,((T=m.data)==null?void 0:T.reload_summary)||"Server reloaded.",!0):N(u,((k=m.data)==null?void 0:k.reason)||"Reload failed",!0)}var y=Yi(),g=l(y),v=p(l(g),2),w=l(v),h=p(w,2),x=l(h),I=p(g,2);{var $=m=>{var T=Vi(),k=l(T);H(()=>A(k,i(u))),E(m,T)};G(I,m=>{i(u)&&m($)})}var C=p(I,2);{var O=m=>{var T=Wi(),k=l(T);H(()=>A(k,i(o))),E(m,T)};G(C,m=>{i(o)&&m(O)})}var P=p(C,2);{var R=m=>{var T=Ji();E(m,T)},q=m=>{var T=Gi();mt(T,21,()=>Object.entries(i(r)),bt,(k,M)=>{var B=Yt(()=>Oa(i(M),2));let D=()=>i(B)[0],L=()=>i(B)[1];var F=Ki(),U=l(F),J=l(U),V=l(J),W=l(V),tt=p(V,2),nt=l(tt),Q=p(J,2),st=l(Q);{var ct=gt=>{An(gt,()=>i(r),D,L)},qt=Yt(()=>typeof L()=="object"&&L()!==null&&!Array.isArray(L())),me=gt=>{Tn(gt,()=>i(r),D,L)};G(st,gt=>{i(qt)?gt(ct):gt(me,-1)})}H(gt=>{A(W,D()),A(nt,gt)},[()=>typeof L()=="object"?Object.keys(L()).length+" keys":typeof L()]),E(k,F)}),E(m,T)};G(P,m=>{i(n)?m(R):i(r)&&m(q,1)})}H(()=>{w.disabled=i(s),h.disabled=i(s),A(x,i(s)?"Saving...":"Save Changes")}),rt("click",w,b),rt("click",h,_),E(t,y),Tt()}be(["change","input","click"]);const $t=(t,e=Ft,r=Ft)=>{var a=Xi(),n=l(a),s=l(n),o=p(n,2),u=l(o);H(()=>{A(s,e()),A(u,r())}),E(t,a)};var Xi=S('<div><div class="text-gray-500"> </div> <div class="text-gray-200 font-mono truncate"> </div></div>'),Qi=S('<p class="text-gray-500">Loading...</p>'),to=S('<tr><td class="px-4 py-2 font-medium"> </td><td class="px-4 py-2 text-gray-400"> </td><td class="px-4 py-2 text-gray-400"> </td><td class="px-4 py-2 text-gray-500 hidden sm:table-cell truncate max-w-32"> </td><td class="px-4 py-2 text-right text-gray-500 hidden md:table-cell tabular-nums"> </td><td class="px-4 py-2 text-right text-gray-500 hidden md:table-cell tabular-nums"> </td><td class="px-4 py-2 text-right text-gray-500 tabular-nums"> </td></tr>'),eo=S('<div class="bg-gray-900 rounded-xl border border-gray-800 p-4 space-y-3"><h3 class="text-sm font-semibold"> </h3> <div class="grid grid-cols-2 sm:grid-cols-3 gap-3 text-xs"><!> <!> <!> <!> <!> <!> <!> <!> <!> <!> <!> <!></div></div>'),ro=S('<div class="bg-gray-900 rounded-xl border border-gray-800 overflow-hidden"><div class="overflow-x-auto"><table class="w-full text-sm"><thead><tr class="text-left text-xs text-gray-500 border-b border-gray-800"><th class="px-4 py-2">Name</th><th class="px-4 py-2">Protocol</th><th class="px-4 py-2">Area</th><th class="px-4 py-2 hidden sm:table-cell">Client</th><th class="px-4 py-2 text-right hidden md:table-cell">Sent</th><th class="px-4 py-2 text-right hidden md:table-cell">Recv</th><th class="px-4 py-2 text-right">Idle</th></tr></thead><tbody class="divide-y divide-gray-800/50"></tbody></table></div></div> <!>',1),ao=S('<div class="space-y-4"><h2 class="text-2xl font-bold">Sessions <span class="text-base text-gray-500 font-normal"> </span></h2> <!></div>');function no(t,e){At(e,!0);let r=j(lt([])),a=j(!0),n=j(null);async function s(){const v=await St("/admin/sessions");v.status===200&&N(r,v.data,!0),N(a,!1)}Zt(()=>{s();const v=setInterval(s,5e3);return()=>clearInterval(v)});function o(v){return v<1024?v+" B":v<1024*1024?(v/1024).toFixed(1)+" KB":(v/(1024*1024)).toFixed(1)+" MB"}var u=ao(),d=l(u),f=p(l(d)),_=l(f),b=p(d,2);{var y=v=>{var w=Qi();E(v,w)},g=v=>{var w=ro(),h=Qe(w),x=l(h),I=l(x),$=p(l(I));mt($,21,()=>i(r),bt,(P,R)=>{var q=to(),m=l(q),T=l(m),k=p(m),M=l(k),B=p(k),D=l(B),L=p(B),F=l(L),U=p(L),J=l(U),V=p(U),W=l(V),tt=p(V),nt=l(tt);H((Q,st)=>{var ct;pt(q,1,`hover:bg-gray-800/30 cursor-pointer ${((ct=i(n))==null?void 0:ct.session_id)===i(R).session_id?"bg-gray-800/50":""}`),A(T,i(R).display_name||"(anonymous)"),A(M,i(R).protocol),A(D,i(R).area),A(F,i(R).client_software),A(J,Q),A(W,st),A(nt,`${i(R).idle_seconds??""}s`)},[()=>o(i(R).bytes_sent),()=>o(i(R).bytes_received)]),rt("click",q,()=>{var Q;return N(n,((Q=i(n))==null?void 0:Q.session_id)===i(R).session_id?null:i(R),!0)}),E(P,q)});var C=p(h,2);{var O=P=>{var R=eo(),q=l(R),m=l(q),T=p(q,2),k=l(T);$t(k,()=>"Session ID",()=>i(n).session_id);var M=p(k,2);$t(M,()=>"Protocol",()=>i(n).protocol);var B=p(M,2);$t(B,()=>"Area",()=>i(n).area);var D=p(B,2);$t(D,()=>"Character",()=>i(n).character_id>=0?"#"+i(n).character_id:"None");var L=p(D,2);$t(L,()=>"HDID",()=>i(n).hardware_id);var F=p(L,2);$t(F,()=>"Client",()=>i(n).client_software);var U=p(F,2);$t(U,()=>"Packets Sent",()=>i(n).packets_sent);var J=p(U,2);$t(J,()=>"Packets Recv",()=>i(n).packets_received);var V=p(J,2);$t(V,()=>"Mod Actions",()=>i(n).mod_actions);var W=p(V,2);{let Q=Yt(()=>o(i(n).bytes_sent));$t(W,()=>"Bytes Sent",()=>i(Q))}var tt=p(W,2);{let Q=Yt(()=>o(i(n).bytes_received));$t(tt,()=>"Bytes Recv",()=>i(Q))}var nt=p(tt,2);$t(nt,()=>"Idle",()=>i(n).idle_seconds+"s"),H(()=>A(m,i(n).display_name||"(anonymous)")),E(P,R)};G(C,P=>{i(n)&&P(O)})}E(v,w)};G(b,v=>{i(a)?v(y):v(g,-1)})}H(()=>A(_,`(${i(r).length??""})`)),E(t,u),Tt()}be(["click"]);var so=S('<span class="inline-block w-2 h-2 rounded-full bg-green-500 ml-2 animate-pulse"></span>'),io=S('<span class="inline-block w-2 h-2 rounded-full bg-red-500 ml-2"></span>'),oo=S('<span class="text-gray-500 shrink-0"> </span>'),lo=S('<div class="px-3 py-1.5 flex gap-2 hover:bg-gray-800/30"><span class="text-gray-600 shrink-0"> </span> <span> </span> <!> <span class="text-yellow-300 shrink-0"> </span> <span class="text-gray-200 break-all"> </span></div>'),fo=S('<div class="px-4 py-8 text-center text-gray-500"> </div>'),uo=S('<div class="space-y-4"><div class="flex items-center justify-between flex-wrap gap-2"><h2 class="text-2xl font-bold">Live Traffic <!></h2> <input type="text" placeholder="Filter by name, text, or area..." class="px-3 py-1.5 text-sm bg-gray-800 border border-gray-700 rounded-lg text-white placeholder-gray-500 w-64"/></div> <div class="bg-gray-900 rounded-xl border border-gray-800 overflow-hidden"><div class="max-h-[70vh] overflow-y-auto divide-y divide-gray-800/50 font-mono text-xs"></div></div></div>');function co(t,e){At(e,!0);let r=j(lt([])),a=j(!1),n=j(""),s=500,o=null;function u(){const C=En();C&&(o=new EventSource("/aonx/v1/events"),o.close(),o=null,d(C))}async function d(C){try{const O=await fetch("/aonx/v1/events",{headers:{Authorization:`Bearer ${C}`}});if(!O.ok){N(a,!1);return}N(a,!0);const P=O.body.getReader(),R=new TextDecoder;let q="";for(;;){const{done:m,value:T}=await P.read();if(m)break;q+=R.decode(T,{stream:!0});const k=q.split(`
+`);q=k.pop()||"";let M="",B="";for(const D of k)D.startsWith("event: ")?M=D.slice(7):D.startsWith("data: ")?B=D.slice(6):D===""&&M&&B&&(f(M,B),M="",B="")}}catch(O){console.error("SSE error:",O)}N(a,!1)}function f(C,O){if(!(C!=="ic_message"&&C!=="ooc_message"))try{const P=JSON.parse(O),R={type:C==="ic_message"?"IC":"OOC",name:P.showname||P.name||P.character||"???",text:P.message||"",time:new Date().toLocaleTimeString(),area:P.area||""};N(r,[R,...i(r).slice(0,s-1)],!0)}catch{}}Zt(()=>(u(),()=>{o&&o.close()}));let _=Yt(()=>i(n)?i(r).filter(C=>C.text.toLowerCase().includes(i(n).toLowerCase())||C.name.toLowerCase().includes(i(n).toLowerCase())||C.area.toLowerCase().includes(i(n).toLowerCase())):i(r));var b=uo(),y=l(b),g=l(y),v=p(l(g));{var w=C=>{var O=so();E(C,O)},h=C=>{var O=io();E(C,O)};G(v,C=>{i(a)?C(w):C(h,-1)})}var x=p(g,2),I=p(y,2),$=l(I);mt($,21,()=>i(_),bt,(C,O)=>{var P=lo(),R=l(P),q=l(R),m=p(R,2),T=l(m),k=p(m,2);{var M=U=>{var J=oo(),V=l(J);H(()=>A(V,`[${i(O).area??""}]`)),E(U,J)};G(k,U=>{i(O).area&&U(M)})}var B=p(k,2),D=l(B),L=p(B,2),F=l(L);H(()=>{A(q,i(O).time),pt(m,1,`shrink-0 w-8 text-center font-bold ${i(O).type==="IC"?"text-blue-400":"text-green-400"}`),A(T,i(O).type),A(D,`${i(O).name??""}:`),A(F,i(O).text)}),E(C,P)},C=>{var O=fo(),P=l(O);H(()=>A(P,i(a)?"Waiting for messages...":"Not connected to event stream")),E(C,O)}),Ye(x,()=>i(n),C=>N(n,C)),E(t,b),Tt()}var vo=S('<p class="text-gray-500">Loading...</p>'),po=S('<tr class="hover:bg-gray-800/30"><td class="px-4 py-2 font-mono text-xs"> </td><td class="px-4 py-2 font-mono text-xs text-gray-500 hidden sm:table-cell"> </td><td class="px-4 py-2 max-w-xs truncate"> </td><td class="px-4 py-2 text-gray-400 hidden md:table-cell"> </td><td class="px-4 py-2 text-gray-500 text-xs hidden lg:table-cell"> </td><td class="px-4 py-2"><span> </span></td><td class="px-4 py-2"><button class="text-xs text-red-400 hover:text-red-300">Unban</button></td></tr>'),ho=S('<tr><td colspan="7" class="px-4 py-8 text-center text-gray-500">No bans found</td></tr>'),_o=S('<div class="bg-gray-900 rounded-xl border border-gray-800 overflow-hidden"><div class="overflow-x-auto"><table class="w-full text-sm"><thead><tr class="text-left text-xs text-gray-500 border-b border-gray-800"><th class="px-4 py-2">IPID</th><th class="px-4 py-2 hidden sm:table-cell">HDID</th><th class="px-4 py-2">Reason</th><th class="px-4 py-2 hidden md:table-cell">Moderator</th><th class="px-4 py-2 hidden lg:table-cell">When</th><th class="px-4 py-2">Duration</th><th class="px-4 py-2"></th></tr></thead><tbody class="divide-y divide-gray-800/50"></tbody></table></div></div>'),go=S('<div class="space-y-4"><div class="flex items-center justify-between flex-wrap gap-2"><h2 class="text-2xl font-bold">Bans</h2> <form class="flex gap-2"><input type="text" placeholder="Search by IPID, HDID, reason..." class="px-3 py-1.5 text-sm bg-gray-800 border border-gray-700 rounded-lg text-white placeholder-gray-500 w-64"/> <button type="submit" class="px-3 py-1.5 text-sm bg-gray-700 hover:bg-gray-600 rounded-lg">Search</button></form></div> <!></div>');function xo(t,e){At(e,!0);let r=j(lt([])),a=j(""),n=j(!0);async function s(){var I;const h=i(a)?`?query=${encodeURIComponent(i(a))}&limit=100`:"?limit=100",x=await St("/admin/bans"+h);x.status===200&&N(r,((I=x.data)==null?void 0:I.bans)||[],!0),N(n,!1)}Zt(()=>{s()});async function o(h){confirm(`Unban ${h}?`)&&(await Ze("/moderation/actions",{action:"unban",target:h}),s())}function u(h){if(h===-2)return"Permanent";if(h===0)return"Invalidated";const x=Math.floor(h/3600),I=Math.floor(h%3600/60);return x>0?`${x}h ${I}m`:`${I}m`}function d(h){return h?new Date(h*1e3).toLocaleString():""}var f=go(),_=l(f),b=p(l(_),2),y=l(b),g=p(_,2);{var v=h=>{var x=vo();E(h,x)},w=h=>{var x=_o(),I=l(x),$=l(I),C=p(l($));mt(C,21,()=>i(r),bt,(O,P)=>{var R=po(),q=l(R),m=l(q),T=p(q),k=l(T),M=p(T),B=l(M),D=p(M),L=l(D),F=p(D),U=l(F),J=p(F),V=l(J),W=l(V),tt=p(J),nt=l(tt);H((Q,st)=>{A(m,i(P).ipid),A(k,i(P).hdid),A(B,i(P).reason),A(L,i(P).moderator),A(U,Q),pt(V,1,`text-xs px-1.5 py-0.5 rounded ${i(P).permanent?"bg-red-900/50 text-red-300":"bg-gray-800 text-gray-400"}`),A(W,st)},[()=>d(i(P).timestamp),()=>u(i(P).duration)]),rt("click",nt,()=>o(i(P).ipid)),E(O,R)},O=>{var P=ho();E(O,P)}),E(h,x)};G(g,h=>{i(n)?h(v):h(w,-1)})}bn("submit",b,h=>{h.preventDefault(),s()}),Ye(y,()=>i(a),h=>N(a,h)),E(t,f),Tt()}be(["click"]);var yo=S('<p class="text-xs mt-2 text-gray-400"> </p>'),bo=S('<p class="text-gray-500">Loading...</p>'),mo=S('<tr class="hover:bg-gray-800/30"><td class="px-3 py-1.5 text-gray-500"> </td><td class="px-3 py-1.5 font-mono"> </td><td class="px-3 py-1.5 text-gray-400"> </td><td class="px-3 py-1.5"><span> </span></td><td class="px-3 py-1.5 text-gray-400 max-w-xs truncate hidden md:table-cell"> </td><td class="px-3 py-1.5 text-gray-500 max-w-xs truncate hidden lg:table-cell"> </td></tr>'),wo=S('<tr><td colspan="6" class="px-4 py-8 text-center text-gray-500">No moderation events</td></tr>'),ko=S('<div class="bg-gray-900 rounded-xl border border-gray-800 overflow-hidden"><div class="overflow-x-auto"><table class="w-full text-xs"><thead><tr class="text-left text-gray-500 border-b border-gray-800"><th class="px-3 py-2">Time</th><th class="px-3 py-2">IPID</th><th class="px-3 py-2">Channel</th><th class="px-3 py-2">Action</th><th class="px-3 py-2 hidden md:table-cell">Message</th><th class="px-3 py-2 hidden lg:table-cell">Reason</th></tr></thead><tbody class="divide-y divide-gray-800/50"></tbody></table></div></div>'),Eo=S('<tr class="hover:bg-gray-800/30"><td class="px-4 py-2 font-mono text-xs"> </td><td class="px-4 py-2 text-gray-400"> </td><td class="px-4 py-2"> </td><td class="px-4 py-2"><button class="text-xs text-red-400 hover:text-red-300">Unmute</button></td></tr>'),So=S('<tr><td colspan="4" class="px-4 py-8 text-center text-gray-500">No active mutes</td></tr>'),Ao=S('<div class="bg-gray-900 rounded-xl border border-gray-800 overflow-hidden"><div class="overflow-x-auto"><table class="w-full text-sm"><thead><tr class="text-left text-xs text-gray-500 border-b border-gray-800"><th class="px-4 py-2">IPID</th><th class="px-4 py-2">Reason</th><th class="px-4 py-2">Remaining</th><th class="px-4 py-2"></th></tr></thead><tbody class="divide-y divide-gray-800/50"></tbody></table></div></div>'),To=S('<div class="space-y-4"><h2 class="text-2xl font-bold">Moderation</h2> <div class="bg-gray-900 rounded-xl border border-gray-800 p-4"><h3 class="text-sm font-semibold text-gray-300 mb-3">Quick Action</h3> <div class="flex flex-wrap gap-2"><input placeholder="IPID or session ID" class="px-3 py-1.5 text-sm bg-gray-800 border border-gray-700 rounded-lg text-white placeholder-gray-500 w-40"/> <input placeholder="Reason (optional)" class="px-3 py-1.5 text-sm bg-gray-800 border border-gray-700 rounded-lg text-white placeholder-gray-500 flex-1 min-w-32"/> <button class="px-3 py-1.5 text-sm bg-yellow-700 hover:bg-yellow-600 rounded-lg">Kick</button> <button class="px-3 py-1.5 text-sm bg-orange-700 hover:bg-orange-600 rounded-lg">Mute</button> <button class="px-3 py-1.5 text-sm bg-red-700 hover:bg-red-600 rounded-lg">Ban</button></div> <!></div> <div class="flex gap-1 border-b border-gray-800"><button>Audit Log</button> <button>Active Mutes</button></div> <!></div>');function Io(t,e){At(e,!0);let r=j("audit"),a=j(lt([])),n=j(lt([])),s=j(!0),o=j(""),u=j(""),d=j("");async function f(){var L;N(s,!0);const D=await St("/admin/moderation-events?limit=100");D.status===200&&N(a,((L=D.data)==null?void 0:L.events)||[],!0),N(s,!1)}async function _(){var L;N(s,!0);const D=await St("/admin/mutes");D.status===200&&N(n,((L=D.data)==null?void 0:L.mutes)||[],!0),N(s,!1)}Zt(()=>{i(r)==="audit"?f():i(r)==="mutes"&&_()});async function b(D){var U;if(!i(o))return;N(d,"");const L={action:D,target:i(o),reason:i(u)||void 0};D==="mute"&&(L.duration=900);const F=await Ze("/moderation/actions",L);N(d,F.status===200?`${D} applied`:((U=F.data)==null?void 0:U.reason)||"Failed",!0),N(o,""),N(u,""),i(r)==="mutes"&&_()}async function y(D){await Ze("/moderation/actions",{action:"unmute",target:D}),_()}var g=To(),v=p(l(g),2),w=p(l(v),2),h=l(w),x=p(h,2),I=p(x,2),$=p(I,2),C=p($,2),O=p(w,2);{var P=D=>{var L=yo(),F=l(L);H(()=>A(F,i(d))),E(D,L)};G(O,D=>{i(d)&&D(P)})}var R=p(v,2),q=l(R),m=p(q,2),T=p(R,2);{var k=D=>{var L=bo();E(D,L)},M=D=>{var L=ko(),F=l(L),U=l(F),J=p(l(U));mt(J,21,()=>i(a),bt,(V,W)=>{var tt=mo(),nt=l(tt),Q=l(nt),st=p(nt),ct=l(st),qt=p(st),me=l(qt),gt=p(qt),gr=l(gt),Dr=l(gr),va=p(gt),In=l(va),Nn=p(va),Cn=l(Nn);H(Mn=>{A(Q,Mn),A(ct,i(W).ipid),A(me,i(W).channel),pt(gr,1,`px-1.5 py-0.5 rounded text-xs
+                    ${i(W).action==="ban"||i(W).action==="perma_ban"?"bg-red-900/50 text-red-300":i(W).action==="kick"?"bg-yellow-900/50 text-yellow-300":i(W).action==="mute"?"bg-orange-900/50 text-orange-300":i(W).action==="censor"?"bg-purple-900/50 text-purple-300":"bg-gray-800 text-gray-400"}`),A(Dr,i(W).action),A(In,i(W).message_sample),A(Cn,i(W).reason)},[()=>new Date(i(W).timestamp_ms).toLocaleString()]),E(V,tt)},V=>{var W=wo();E(V,W)}),E(D,L)},B=D=>{var L=Ao(),F=l(L),U=l(F),J=p(l(U));mt(J,21,()=>i(n),bt,(V,W)=>{var tt=Eo(),nt=l(tt),Q=l(nt),st=p(nt),ct=l(st),qt=p(st),me=l(qt),gt=p(qt),gr=l(gt);H(Dr=>{A(Q,i(W).ipid),A(ct,i(W).reason),A(me,Dr)},[()=>i(W).seconds_remaining<0?"Permanent":Math.ceil(i(W).seconds_remaining/60)+"m"]),rt("click",gr,()=>y(i(W).ipid)),E(V,tt)},V=>{var W=So();E(V,W)}),E(D,L)};G(T,D=>{i(s)?D(k):i(r)==="audit"?D(M,1):i(r)==="mutes"&&D(B,2)})}H(()=>{pt(q,1,`px-4 py-2 text-sm ${i(r)==="audit"?"text-white border-b-2 border-blue-500":"text-gray-400"}`),pt(m,1,`px-4 py-2 text-sm ${i(r)==="mutes"?"text-white border-b-2 border-blue-500":"text-gray-400"}`)}),Ye(h,()=>i(o),D=>N(o,D)),Ye(x,()=>i(u),D=>N(u,D)),rt("click",I,()=>b("kick")),rt("click",$,()=>b("mute")),rt("click",C,()=>b("ban")),rt("click",q,()=>N(r,"audit")),rt("click",m,()=>N(r,"mutes")),E(t,g),Tt()}be(["click"]);var No=S('<p class="text-gray-500">Loading...</p>'),Co=S('<span class="px-1.5 py-0.5 rounded bg-blue-900/50 text-blue-300"> </span>'),Mo=S('<span class="px-1.5 py-0.5 rounded bg-gray-800 text-gray-500">IDLE</span>'),$o=S('<span class="px-1.5 py-0.5 rounded bg-yellow-900/50 text-yellow-300"> </span>'),Ro=S('<span class="px-1.5 py-0.5 rounded bg-green-900/50 text-green-300"> </span>'),Do=S('<button><div class="flex items-center justify-between mb-2"><span class="font-semibold text-sm"> </span> <span class="text-xs text-gray-500"> </span></div> <div class="flex gap-2 text-xs"><!> <!> <!></div></button>'),Oo=S('<div><span class="text-gray-500">Background</span><br/> </div>'),Lo=S('<div><span class="text-gray-500">Music</span><br/> </div>'),Po=S('<div><span class="text-gray-500">HP</span><br/> </div>'),Fo=S('<div class="bg-gray-900 rounded-xl border border-gray-800 p-4 space-y-3"><h3 class="text-sm font-semibold"> </h3> <div class="grid grid-cols-2 sm:grid-cols-3 gap-3 text-xs"><!> <!> <!></div></div>'),jo=S('<div class="grid gap-3 sm:grid-cols-2 lg:grid-cols-3"></div> <!>',1),Bo=S('<div class="space-y-4"><h2 class="text-2xl font-bold">Areas</h2> <!></div>');function Uo(t,e){At(e,!0);let r=j(lt([])),a=j(!0),n=j(null);async function s(){var y;const b=await St("/areas");b.status===200&&N(r,((y=b.data)==null?void 0:y.areas)||[],!0),N(a,!1)}Zt(()=>{s();const b=setInterval(s,1e4);return()=>clearInterval(b)});async function o(b){const y=await St(`/areas/${b}`);y.status===200&&N(n,y.data,!0)}var u=Bo(),d=p(l(u),2);{var f=b=>{var y=No();E(b,y)},_=b=>{var y=jo(),g=Qe(y);mt(g,21,()=>i(r),bt,(h,x)=>{var I=Do(),$=l(I),C=l($),O=l(C),P=p(C,2),R=l(P),q=p($,2),m=l(q);{var T=F=>{var U=Co(),J=l(U);H(()=>A(J,i(x).status)),E(F,U)},k=F=>{var U=Mo();E(F,U)};G(m,F=>{i(x).status&&i(x).status!=="IDLE"?F(T):F(k,-1)})}var M=p(m,2);{var B=F=>{var U=$o(),J=l(U);H(()=>A(J,i(x).locked)),E(F,U)};G(M,F=>{i(x).locked&&i(x).locked!=="FREE"&&F(B)})}var D=p(M,2);{var L=F=>{var U=Ro(),J=l(U);H(()=>A(J,`CM: ${i(x).cm??""}`)),E(F,U)};G(D,F=>{i(x).cm&&F(L)})}H(()=>{var F,U;pt(I,1,`bg-gray-900 rounded-xl border border-gray-800 p-4 text-left hover:border-gray-600 transition-colors
+                 ${((U=(F=i(n))==null?void 0:F.area)==null?void 0:U.id)===i(x).id?"border-blue-500":""}`),A(O,i(x).name),A(R,`${i(x).players??0??""} players`)}),rt("click",I,()=>o(i(x).id)),E(h,I)});var v=p(g,2);{var w=h=>{var x=Fo(),I=l(x),$=l(I),C=p(I,2),O=l(C);{var P=k=>{var M=Oo(),B=p(l(M),2,!0);H(()=>A(B,i(n).background.name)),E(k,M)};G(O,k=>{i(n).background&&k(P)})}var R=p(O,2);{var q=k=>{var M=Lo(),B=p(l(M),2,!0);H(()=>A(B,i(n).music.name||"None")),E(k,M)};G(R,k=>{i(n).music&&k(q)})}var m=p(R,2);{var T=k=>{var M=Po(),B=p(l(M),2);H(()=>A(B,`Def: ${i(n).hp.defense??""} / Pro: ${i(n).hp.prosecution??""}`)),E(k,M)};G(m,k=>{i(n).hp&&k(T)})}H(()=>{var k;return A($,((k=i(n).area)==null?void 0:k.name)||"Area Detail")}),E(h,x)};G(v,h=>{i(n)&&h(w)})}E(b,y)};G(d,b=>{i(a)?b(f):b(_,-1)})}E(t,u),Tt()}be(["click"]);var zo=S('<p class="text-gray-500">Loading...</p>'),Ho=S('<tr class="hover:bg-gray-800/30"><td class="px-4 py-2 font-medium"> </td><td class="px-4 py-2"><span> </span></td></tr>'),qo=S('<tr><td colspan="2" class="px-4 py-8 text-center text-gray-500">No users</td></tr>'),Vo=S('<div class="bg-gray-900 rounded-xl border border-gray-800 overflow-hidden"><div class="overflow-x-auto"><table class="w-full text-sm"><thead><tr class="text-left text-xs text-gray-500 border-b border-gray-800"><th class="px-4 py-2">Username</th><th class="px-4 py-2">Role</th></tr></thead><tbody class="divide-y divide-gray-800/50"></tbody></table></div></div>'),Wo=S('<div class="space-y-4"><h2 class="text-2xl font-bold">Users</h2> <!></div>');function Jo(t,e){At(e,!0);let r=j(lt([])),a=j(!0);async function n(){var _;const f=await St("/admin/users");f.status===200&&N(r,((_=f.data)==null?void 0:_.users)||[],!0),N(a,!1)}Zt(()=>{n()});var s=Wo(),o=p(l(s),2);{var u=f=>{var _=zo();E(f,_)},d=f=>{var _=Vo(),b=l(_),y=l(b),g=p(l(y));mt(g,21,()=>i(r),bt,(v,w)=>{var h=Ho(),x=l(h),I=l(x),$=p(x),C=l($),O=l(C);H(()=>{A(I,i(w).username),pt(C,1,`text-xs px-1.5 py-0.5 rounded
+                    ${i(w).acl==="SUPER"?"bg-purple-900/50 text-purple-300":i(w).acl==="NONE"?"bg-gray-800 text-gray-500":"bg-blue-900/50 text-blue-300"}`),A(O,i(w).acl)}),E(v,h)},v=>{var w=qo();E(v,w)}),E(f,_)};G(o,f=>{i(a)?f(u):f(d,-1)})}E(t,s),Tt()}var Ko=S('<p class="text-gray-500">Loading...</p>'),Go=S('<tr class="hover:bg-gray-800/30"><td class="px-3 py-2 font-mono text-xs"> </td><td class="px-3 py-2 text-gray-400 max-w-xs truncate"> </td><td class="px-3 py-2 text-gray-500 text-xs hidden md:table-cell"> </td><td class="px-3 py-2 text-xs"> </td></tr>'),Yo=S('<table class="w-full text-sm"><thead><tr class="text-left text-xs text-gray-500 border-b border-gray-800"><th class="px-3 py-2">Target</th><th class="px-3 py-2">Reason</th><th class="px-3 py-2 hidden md:table-cell">Installed</th><th class="px-3 py-2">Expires</th></tr></thead><tbody class="divide-y divide-gray-800/50"></tbody></table>'),Zo=S('<p class="text-sm text-gray-500">No active firewall rules.</p>'),Xo=S('<div class="bg-gray-900 rounded-xl border border-gray-800 p-4"><div class="flex items-center gap-2 mb-3"><span class="text-sm font-semibold">nftables</span> <span> </span> <span class="text-xs text-gray-500"> </span></div> <!></div>'),Qo=S('<tr class="hover:bg-gray-800/30"><td class="px-4 py-2 font-mono text-xs"> </td><td class="px-4 py-2 text-gray-300"> </td><td class="px-4 py-2"><span> </span></td><td class="px-4 py-2 text-gray-500 hidden md:table-cell"> </td><td class="px-4 py-2 text-gray-500 hidden md:table-cell"> </td><td class="px-4 py-2 text-gray-500 text-xs max-w-xs truncate hidden lg:table-cell"> </td></tr>'),tl=S('<tr><td colspan="6" class="px-4 py-8 text-center text-gray-500">No flagged ASNs</td></tr>'),el=S('<div class="bg-gray-900 rounded-xl border border-gray-800 overflow-hidden"><div class="overflow-x-auto"><table class="w-full text-sm"><thead><tr class="text-left text-xs text-gray-500 border-b border-gray-800"><th class="px-4 py-2">ASN</th><th class="px-4 py-2">Organization</th><th class="px-4 py-2">Status</th><th class="px-4 py-2 hidden md:table-cell">Abuse Events</th><th class="px-4 py-2 hidden md:table-cell">Abusive IPs</th><th class="px-4 py-2 hidden lg:table-cell">Block Reason</th></tr></thead><tbody class="divide-y divide-gray-800/50"></tbody></table></div></div>'),rl=S('<div class="space-y-4"><h2 class="text-2xl font-bold">Firewall &amp; Network</h2> <div class="flex gap-1 border-b border-gray-800"><button>IP Rules</button> <button>ASN Reputation</button></div> <!></div>');function al(t,e){At(e,!0);let r=j("firewall"),a=j(lt({enabled:!1,rules:[]})),n=j(lt([])),s=j(!0);async function o(){N(s,!0);const x=await St("/admin/firewall");x.status===200&&N(a,x.data,!0),N(s,!1)}async function u(){var I;N(s,!0);const x=await St("/admin/asn-reputation");x.status===200&&N(n,((I=x.data)==null?void 0:I.asn_entries)||[],!0),N(s,!1)}Zt(()=>{i(r)==="firewall"?o():u()});function d(x){return x?new Date(x*1e3).toLocaleString():"Never"}var f=rl(),_=p(l(f),2),b=l(_),y=p(b,2),g=p(_,2);{var v=x=>{var I=Ko();E(x,I)},w=x=>{var I=Xo(),$=l(I),C=p(l($),2),O=l(C),P=p(C,2),R=l(P),q=p($,2);{var m=k=>{var M=Yo(),B=p(l(M));mt(B,21,()=>i(a).rules,bt,(D,L)=>{var F=Go(),U=l(F),J=l(U),V=p(U),W=l(V),tt=p(V),nt=l(tt),Q=p(tt),st=l(Q);H((ct,qt)=>{A(J,i(L).target),A(W,i(L).reason),A(nt,ct),A(st,qt)},[()=>d(i(L).installed_at),()=>i(L).expires_at===0?"Permanent":d(i(L).expires_at)]),E(D,F)}),E(k,M)},T=k=>{var M=Zo();E(k,M)};G(q,k=>{i(a).rules.length>0?k(m):k(T,-1)})}H(()=>{pt(C,1,`text-xs px-1.5 py-0.5 rounded ${i(a).enabled?"bg-green-900/50 text-green-300":"bg-gray-800 text-gray-500"}`),A(O,i(a).enabled?"Enabled":"Disabled"),A(R,`${i(a).rules.length??""} rule(s)`)}),E(x,I)},h=x=>{var I=el(),$=l(I),C=l($),O=p(l(C));mt(O,21,()=>i(n),bt,(P,R)=>{var q=Qo(),m=l(q),T=l(m),k=p(m),M=l(k),B=p(k),D=l(B),L=l(D),F=p(B),U=l(F),J=p(F),V=l(J),W=p(J),tt=l(W);H(()=>{A(T,`AS${i(R).asn??""}`),A(M,i(R).as_org),pt(D,1,`text-xs px-1.5 py-0.5 rounded
+                    ${i(R).status==="blocked"?"bg-red-900/50 text-red-300":i(R).status==="rate_limited"?"bg-orange-900/50 text-orange-300":i(R).status==="watched"?"bg-yellow-900/50 text-yellow-300":"bg-gray-800 text-gray-400"}`),A(L,i(R).status),A(U,i(R).total_abuse_events),A(V,i(R).abusive_ips),A(tt,i(R).block_reason)}),E(P,q)},P=>{var R=tl();E(P,R)}),E(x,I)};G(g,x=>{i(s)?x(v):i(r)==="firewall"?x(w,1):x(h,-1)})}H(()=>{pt(b,1,`px-4 py-2 text-sm ${i(r)==="firewall"?"text-white border-b-2 border-blue-500":"text-gray-400"}`),pt(y,1,`px-4 py-2 text-sm ${i(r)==="asn"?"text-white border-b-2 border-blue-500":"text-gray-400"}`)}),rt("click",b,()=>N(r,"firewall")),rt("click",y,()=>N(r,"asn")),E(t,f),Tt()}be(["click"]);var nl=S('<p class="text-gray-500">Loading...</p>'),sl=S('<div class="px-4 py-2 flex items-center gap-3 text-sm hover:bg-gray-800/30"><span class="text-xs text-gray-600 w-8 text-right tabular-nums"></span> <span class="text-gray-200"> </span></div>'),il=S('<div class="px-4 py-8 text-center text-gray-500">No items</div>'),ol=S('<div class="bg-gray-900 rounded-xl border border-gray-800 overflow-hidden"><div class="max-h-[60vh] overflow-y-auto divide-y divide-gray-800/50"></div></div>'),ll=S('<div class="space-y-4"><h2 class="text-2xl font-bold">Content</h2> <div class="flex gap-1 border-b border-gray-800"><button> </button> <button> </button> <button> </button></div> <!></div>');function dl(t,e){At(e,!0);let r=j("areas"),a=j(lt({characters:[],music:[],areas:[]})),n=j(!0);async function s(){const I=await St("/admin/content");I.status===200&&N(a,I.data,!0),N(n,!1)}Zt(()=>{s()});function o(){return i(r)==="areas"?i(a).areas:i(r)==="characters"?i(a).characters:i(r)==="music"?i(a).music:[]}var u=ll(),d=p(l(u),2),f=l(d),_=l(f),b=p(f,2),y=l(b),g=p(b,2),v=l(g),w=p(d,2);{var h=I=>{var $=nl();E(I,$)},x=I=>{var $=ol(),C=l($);mt(C,21,o,bt,(O,P,R)=>{var q=sl(),m=l(q);m.textContent=R+1;var T=p(m,2),k=l(T);H(()=>A(k,i(P))),E(O,q)},O=>{var P=il();E(O,P)}),E(I,$)};G(w,I=>{i(n)?I(h):I(x,-1)})}H(()=>{pt(f,1,`px-4 py-2 text-sm ${i(r)==="areas"?"text-white border-b-2 border-blue-500":"text-gray-400"}`),A(_,`Areas (${i(a).areas.length??""})`),pt(b,1,`px-4 py-2 text-sm ${i(r)==="characters"?"text-white border-b-2 border-blue-500":"text-gray-400"}`),A(y,`Characters (${i(a).characters.length??""})`),pt(g,1,`px-4 py-2 text-sm ${i(r)==="music"?"text-white border-b-2 border-blue-500":"text-gray-400"}`),A(v,`Music (${i(a).music.length??""})`)}),rt("click",f,()=>N(r,"areas")),rt("click",b,()=>N(r,"characters")),rt("click",g,()=>N(r,"music")),E(t,u),Tt()}be(["click"]);var fl=S('<div class="text-center text-gray-500 mt-20"><h2 class="text-2xl font-semibold mb-2">Coming Soon</h2> <p>The <code class="text-gray-400"> </code> page is not yet implemented.</p></div>'),ul=S('<div class="flex h-screen bg-gray-950 text-gray-100"><!> <main class="flex-1 overflow-auto p-4 md:p-6 lg:p-8"><!></main></div>');function cl(t,e){At(e,!0);let r=j(lt(window.location.hash||"#/login"));function a(){N(r,window.location.hash||"#/login",!0)}Zt(()=>(window.addEventListener("hashchange",a),()=>window.removeEventListener("hashchange",a)));let n=Yt(()=>!it.loggedIn&&i(r)!=="#/login"?(window.location.hash="#/login","login"):it.loggedIn&&i(r)==="#/login"?(window.location.hash="#/dashboard","dashboard"):i(r).slice(2)||"login");var s=kn(),o=Qe(s);{var u=f=>{Ei(f,{})},d=f=>{var _=ul(),b=l(_);mi(b,{get currentPage(){return i(n)}});var y=p(b,2),g=l(y);{var v=m=>{Pi(m,{})},w=m=>{Zi(m,{})},h=m=>{no(m,{})},x=m=>{co(m,{})},I=m=>{xo(m,{})},$=m=>{Io(m,{})},C=m=>{Uo(m,{})},O=m=>{Jo(m,{})},P=m=>{al(m,{})},R=m=>{dl(m,{})},q=m=>{var T=fl(),k=p(l(T),2),M=p(l(k)),B=l(M);H(()=>A(B,i(n))),E(m,T)};G(g,m=>{i(n)==="dashboard"?m(v):i(n)==="config"?m(w,1):i(n)==="sessions"?m(h,2):i(n)==="traffic"?m(x,3):i(n)==="bans"?m(I,4):i(n)==="moderation"?m($,5):i(n)==="areas"?m(C,6):i(n)==="users"?m(O,7):i(n)==="firewall"?m(P,8):i(n)==="content"?m(R,9):m(q,-1)})}E(f,_)};G(o,f=>{it.loggedIn?f(d,-1):f(u)})}E(t,s),Tt()}Zs(cl,{target:document.getElementById("app")});
