+++ assets/admin/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Kagami Admin</title>
-    <script type="module" crossorigin src="/admin/assets/index-CKJRbsUW.js"></script>
+    <script type="module" crossorigin src="/admin/assets/index-BabJX8K1.js"></script>
     <link rel="stylesheet" crossorigin href="/admin/assets/index-CZeh9xlq.css">
   </head>
   <body>
